### 2025-12-01 26cdeff
ci: align pr checks and release-please with ADR-009
