+++ .github/workflows/release-please.yaml
@@ -1,0 +1,32 @@
+name: release-please
+on:
+  push:
+    branches: [ develop ]
+  workflow_dispatch:
+
+permissions:
+  contents: write
+  pull-requests: write
+
+jobs:
+  release-please:
+    runs-on: ubuntu-latest
+    steps:
+      - name: release-please
+        uses: googleapis/release-please-action@v4
+        with:
+          release-type: simple
+          package-name: bloch
+          changelog-types: >
+            [{"type":"feat","section":"Features","hidden":false},
+             {"type":"fix","section":"Bug Fixes","hidden":false},
+             {"type":"perf","section":"Performance","hidden":false},
+             {"type":"refactor","section":"Refactoring","hidden":false},
+             {"type":"revert","section":"Reverts","hidden":false},
+             {"type":"docs","section":"Docs","hidden":false},
+             {"type":"test","section":"Tests","hidden":false},
+             {"type":"build","section":"Build System","hidden":false},
+             {"type":"ci","section":"CI/CD","hidden":false},
+             {"type":"chore","section":"Chores","hidden":false}]
+          bump-minor-pre-major: true
+          bump-patch-for-minor-pre-major: true
