### 2025-12-01 63e11c2
ci: align pr checks and release-please with ADR-009
