+++ .github/workflows/pr-checks.yaml
@@ -2,7 +2,7 @@ name: PR Checks
 
 on:
   pull_request:
-    branches: [ master ]
+    branches: [develop, master]
 
 permissions:
   contents: read
