### 2025-12-01 5e2a709
ci: align pr checks and release-please with ADR-009
