+++ .github/workflows/static-analysis.yaml
@@ -1,49 +1,0 @@
-name: Static Analysis
-
-on:
-  pull_request:
-    branches:
-      - develop
-      - master
-      - 'release-v*'
-  schedule:
-    - cron: '0 3 * * 1'
-
-permissions:
-  actions: read
-  contents: read
-  security-events: write
-
-jobs:
-  analyze:
-    name: CodeQL analyze
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        language: ['cpp']
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Install build dependencies
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y cmake g++
-
-      - name: Initialize CodeQL
-        uses: github/codeql-action/init@v3
-        with:
-          languages: ${{ matrix.language }}
-          queries: security-and-quality
-
-      - name: Configure build
-        run: cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
-
-      - name: Build project
-        run: cmake --build build --parallel
-
-      - name: Perform CodeQL Analysis
-        uses: github/codeql-action/analyze@v3
