+++ .github/workflows/pr-checks.yaml
@@ -2,7 +2,9 @@ name: PR Checks
 
 on:
   pull_request:
-    branches: [develop, master]
+    branches:
+      - develop
+      - master
 
 permissions:
   contents: read
