### 2025-12-01 ba89862
Revert "ci: align pr checks and release-please with ADR-009 (#160)"
This reverts commit a9fa888c31462d1a84c94134c273398e38ae4343.
