### 2025-12-01 e68d783
Revert "ci: align pr checks and release-please with ADR-009" (#161)
