### 2025-12-01 86f7b2c
ci: align release-please and pr-checks workflows with ADR-009 (#162)
