+++ .github/ISSUE_TEMPLATE/defect.md
@@ -1,0 +1,20 @@
+---
+name: Defect
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Problem Overview**
+A clear and concise description of what the bug is.
+
+**Impact**
+How does this impact behaviour
+
+**Steps to Recreate**
+Any steps required to recreate the defect locally
+
+**Fix Ideas**
+Any ideas what to change
