### 2025-12-01 3fc46c7
fix(perf): made runtime evaluator single use to prevent state leakage
