### 2025-12-01 87afa42
fix(perf): made runtime evaluator single use to prevent state leakage (#170)
