### 2025-12-01 9255007
ci: fix release-please workflow
