+++ .github/workflows/release-please.yaml
@@ -17,6 +17,7 @@ jobs:
         with:
           release-type: simple
           package-name: bloch
+          default-branch: develop
           changelog-types: >
             [{"type":"feat","section":"Features","hidden":false},
              {"type":"fix","section":"Bug Fixes","hidden":false},
