### 2025-12-01 0cdbbbb
hotfix: fix release-please workflow
