+++ .github/workflows/release-please.yaml
@@ -18,6 +18,7 @@ jobs:
           release-type: simple
           package-name: bloch
           default-branch: develop
+          target-branch: develop
           changelog-types: >
             [{"type":"feat","section":"Features","hidden":false},
              {"type":"fix","section":"Bug Fixes","hidden":false},
