### 2025-12-01 7d74ad8
fix: reset now clears measured flag
