### 2025-12-01 e07f9fd
fix: reset now clears measured flag (#172)
