### 2025-12-09 21f4b68
fix: initialise m_state to a single amplitude
