### 2025-12-09 b6d59e2
fix: initialise m_state to a single amplitude (#177)
