### 2025-12-09 a941ac4
fix: optimise qasm logging for better performance
