### 2025-12-09 99c0ba1
fix: optimise qasm logging for better performance
