### 2025-12-09 de18ef0
fix: optimise qasm logging for better performance
