### 2025-12-09 a2b9cd4
fix: optimise qasm logging for better performance (#178)
