### 2025-12-09 5db209b
fix: cx gate works on subspace where ctrl is 1 only
