### 2025-12-09 6938c5d
fix: cx gate works on subspace where ctrl is 1 only (#180)
