### 2025-12-09 68c6371
fix: parenthesised expressions not being evaluated at runtime (#182)
