### 2025-12-11 7bbaedd
fix: install scripts migrated to website, added update checker
