### 2025-12-11 fee4115
fix: install scripts migrated to website, added update checker (#184)
