+++ src/bloch/update/update_manager.cpp
@@ -1,0 +1,674 @@
+// Copyright 2025 Akshay Pal (https://bloch-labs.com)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "bloch/update/update_manager.hpp"
+
+#include <cctype>
+#include <chrono>
+#include <cstdlib>
+#include <filesystem>
+#include <fstream>
+#include <iomanip>
+#include <iostream>
+#include <memory>
+#include <optional>
+#include <sstream>
+#include <string>
+#include <string_view>
+#include <utility>
+
+#ifndef CPPHTTPLIB_OPENSSL_SUPPORT
+#define CPPHTTPLIB_OPENSSL_SUPPORT
+#endif
+#include <openssl/evp.h>
+#include "third_party/cpp-httplib/httplib.h"
+
+namespace bloch {
+    namespace {
+
+        using Clock = std::chrono::system_clock;
+        constexpr auto kUpdateWindow = std::chrono::hours(72);
+        constexpr std::string_view kChangelogUrl =
+            "https://github.com/bloch-labs/bloch/blob/master/CHANGELOG.md";
+
+        struct UpdateCache {
+            std::string latestVersion;
+            Clock::time_point lastChecked{};
+            Clock::time_point lastNotified{};
+        };
+
+        struct SemVer {
+            int major{0};
+            int minor{0};
+            int patch{0};
+            bool valid{false};
+        };
+
+        std::string userAgent(const std::string& currentVersion) {
+            if (currentVersion.empty())
+                return "bloch";
+            return "bloch/" + currentVersion;
+        }
+
+        std::filesystem::path cacheFilePath() {
+            if (const char* xdg = std::getenv("XDG_CACHE_HOME"); xdg && *xdg) {
+                return std::filesystem::path(xdg) / "bloch" / "update_cache.txt";
+            }
+#ifdef _WIN32
+            if (const char* local = std::getenv("LOCALAPPDATA"); local && *local) {
+                return std::filesystem::path(local) / "Bloch" / "update_cache.txt";
+            }
+#endif
+            if (const char* home = std::getenv("HOME"); home && *home) {
+                return std::filesystem::path(home) / ".cache" / "bloch" / "update_cache.txt";
+            }
+            return std::filesystem::temp_directory_path() / "bloch-update-cache.txt";
+        }
+
+        UpdateCache emptyCache() {
+            return UpdateCache{std::string(), Clock::time_point{}, Clock::time_point{}};
+        }
+
+        std::optional<UpdateCache> loadCache() {
+            const auto path = cacheFilePath();
+            std::ifstream in(path);
+            if (!in)
+                return std::nullopt;
+
+            std::string checkedStr;
+            std::string latest;
+            std::string notifiedStr;
+            if (!std::getline(in, checkedStr))
+                return std::nullopt;
+            if (!std::getline(in, latest))
+                return std::nullopt;
+            if (!std::getline(in, notifiedStr))
+                notifiedStr = "0";
+
+            try {
+                auto checkedSec = std::stoll(checkedStr);
+                auto notifiedSec = std::stoll(notifiedStr.empty() ? "0" : notifiedStr);
+                UpdateCache cache;
+                cache.latestVersion = latest;
[…]
