+++ src/bloch/update/update_manager.hpp
@@ -1,0 +1,26 @@
+// Copyright 2025 Akshay Pal (https://bloch-labs.com)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <string>
+
+namespace bloch {
+    // Run a lightweight update check (no more than once every 72 hours) and
+    // print a single notice if a newer version exists.
+    void checkForUpdatesIfDue(const std::string& currentVersion);
+
+    // Download and install the latest Bloch release. Returns true on success.
+    bool performSelfUpdate(const std::string& currentVersion, const std::string& argv0);
+}  // namespace bloch
