+++ scripts/install.sh
@@ -1,299 +1,0 @@
-#!/usr/bin/env bash
-
-if [ -z "${BASH_VERSION:-}" ]; then
-  echo "This installer requires bash. Re-run with: curl -fsSL https://raw.githubusercontent.com/bloch-labs/bloch/HEAD/scripts/install.sh | bash -s -- <version>" >&2
-  exit 1
-fi
-
-set -euo pipefail
-
-# Pretty output helpers
-if [[ -t 1 && "${NO_COLOR:-}" != "1" ]]; then
-  RED="\033[31m"; GREEN="\033[32m"; YELLOW="\033[33m"; BLUE="\033[34m"; MAGENTA="\033[35m"; CYAN="\033[36m"; BOLD="\033[1m"; DIM="\033[2m"; RESET="\033[0m"
-else
-  RED=""; GREEN=""; YELLOW=""; BLUE=""; MAGENTA=""; CYAN=""; BOLD=""; DIM=""; RESET=""
-fi
-
-step()    { printf "%b➜%b %s\n" "$BLUE" "$RESET" "$*"; }
-success() { printf "%b✓%b %s\n" "$GREEN" "$RESET" "$*"; }
-warn()    { printf "%b!%b %s\n" "$YELLOW" "$RESET" "$*" >&2; }
-error()   { printf "%b✗%b %s\n" "$RED" "$RESET" "$*" >&2; }
-note()    { printf "%b%s%b\n" "$DIM" "$*" "$RESET"; }
-
-has_tty() {
-  # Consider interactive if stdout or stderr is a TTY, or FORCE_INTERACTIVE is set
-  [[ -t 1 || -t 2 || -n "${FORCE_INTERACTIVE:-}" ]]
-}
-
-# Bloch installer for Linux/macOS
-# Usage examples:
-#   curl -fsSL https://raw.githubusercontent.com/bloch-labs/bloch/HEAD/scripts/install.sh | sh -s -- v1.0.0-rc.1
-#   curl -fsSL https://raw.githubusercontent.com/bloch-labs/bloch/HEAD/scripts/install.sh | sh -s -- latest
-#   INSTALL_DIR=$HOME/.local/bin sh scripts/install.sh v1.0.0-rc.1
-
-REPO="bloch-labs/bloch"
-
-usage() {
-  cat <<EOF
-Bloch installer
-
-Usage:
-  $0 <version|latest|latest-rc>
-
-Environment variables:
-  INSTALL_DIR   Destination directory (default: /usr/local/bin or ~/.local/bin fallback)
-
-Examples:
-  $0 v1.0.0-rc.1
-  $0 latest
-  INSTALL_DIR=\$HOME/.local/bin $0 v1.0.0-rc.1
-EOF
-}
-
-if [[ "${1:-}" == "-h" || "${1:-}" == "--help" ]]; then
-  usage
-  exit 0
-fi
-
-VERSION_INPUT=${1:-latest}
-
-require_cmd() {
-  command -v "$1" >/dev/null 2>&1 || { echo "Error: required command '$1' not found" >&2; exit 1; }
-}
-
-require_cmd curl
-require_cmd tar
-
-# Detect OS
-UNAME_S=$(uname -s)
-case "$UNAME_S" in
-  Linux) OS="Linux" ;;
-  Darwin) OS="macOS" ;;
-  *) echo "Error: unsupported OS: $UNAME_S" >&2; exit 1 ;;
-esac
-
-# Detect ARCH mapping to GitHub runner arch naming
-UNAME_M=$(uname -m)
-case "$UNAME_M" in
-  x86_64|amd64) ARCH="X64" ;;
-  arm64|aarch64) ARCH="ARM64" ;;
[…]
