### 2025-12-11 e4e9d58
fix: fixed bug where arrays could only be initialised by literals
