### 2025-12-11 4c676f3
fix: fixed bug where arrays could only be initialised by literals (#185)
