### 2025-12-11 7a5197b
ci: hotfix the manual packager (#186)
