+++ .github/workflows/manual-packager.yaml
@@ -18,6 +18,8 @@ jobs:
   package:
     env:
       VCPKG_ROOT: ${{ github.workspace }}/.vcpkg
+      VCPKG_FORCE_SYSTEM_BINARIES: 1
+      VCPKG_DOWNLOADS: ${{ github.workspace }}/.vcpkg_downloads
     strategy:
       matrix:
         include:
@@ -26,12 +28,12 @@ jobs:
             arch: X64
             cpack_generator: TGZ
             artifact_ext: tar.gz
-          - runner: macos-13
+          - runner: macos-15-intel
             os_name: macOS
             arch: X64
             cpack_generator: TGZ
             artifact_ext: tar.gz
-          - runner: macos-14
+          - runner: macos-latest
             os_name: macOS
             arch: ARM64
             cpack_generator: TGZ
@@ -48,6 +50,17 @@ jobs:
         uses: actions/checkout@v4
         with:
           ref: ${{ inputs.branch }}
+
+      - name: Ensure CMake 3.31.10 (Windows)
+        if: matrix.runner == 'windows-latest'
+        shell: pwsh
+        run: |
+          $ver = "3.31.10"
+          $zip = "$env:RUNNER_TEMP\cmake-$ver.zip"
+          Invoke-WebRequest -Uri "https://cmake.org/files/v3.31/cmake-$ver-windows-x86_64.zip" -OutFile $zip
+          Expand-Archive $zip -DestinationPath "$env:RUNNER_TEMP\cmake"
+          $cmakeDir = Get-ChildItem "$env:RUNNER_TEMP\cmake" -Directory | Select-Object -First 1
+          "$($cmakeDir.FullName)\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
 
       - name: Install build deps (Linux)
         if: matrix.runner == 'ubuntu-latest'
