+++ .github/workflows/release-please.yaml
@@ -1,7 +1,7 @@
 name: release-please
 on:
   push:
-    branches: develop
+    branches: master
   workflow_dispatch:
 
 permissions:
@@ -17,8 +17,8 @@ jobs:
         with:
           release-type: simple
           package-name: bloch
-          default-branch: develop
-          target-branch: develop
+          default-branch: master
+          target-branch: master
           changelog-types: >
             [{"type":"feat","section":"Features","hidden":false},
              {"type":"fix","section":"Bug Fixes","hidden":false},
