### 2025-12-11 58e0472
chore(master): release 1.0.2
