### 2025-12-11 7b4a8db
ci: hotfix packager
