### 2025-12-11 000181e
ci: hotfix packager
