### 2025-12-11 10aba66
ci: hotfix packager (#192)
