### 2025-12-11 8358913
Reinstate install.sh
