### 2025-12-12 #9503 @bloch-akshay
@danielskatz I think the following would be appropritate

@HectorMozo3110
@dorchard
@jdoerfort
@kareem1925
@joshiCoding

As they combine knowledge of quantum computing and programming language theory.
