### 2025-12-13 c859a99
build: reorg app arch
