### 2025-12-13 d08765c
build: reorg app arch
