### 2025-12-13 afd8b3f
build: reorg app arch (#195)
