+++ tests/test_runtime.cpp
@@ -13,14 +13,18 @@
 // limitations under the License.
 
 #include <sstream>
-#include "bloch/lexer/lexer.hpp"
-#include "bloch/parser/parser.hpp"
+
+#include "bloch/core/lexer/lexer.hpp"
+#include "bloch/core/parser/parser.hpp"
+#include "bloch/core/semantics/semantic_analyser.hpp"
 #include "bloch/runtime/qasm_simulator.hpp"
 #include "bloch/runtime/runtime_evaluator.hpp"
-#include "bloch/semantics/semantic_analyser.hpp"
+#include "bloch/support/error/bloch_error.hpp"
 #include "test_framework.hpp"
 
-using namespace bloch;
+using namespace bloch::core;
+using namespace bloch::runtime;
+using bloch::support::BlochError;
 
 static std::unique_ptr<Program> parseProgram(const char* src) {
     Lexer lexer(src);
