### 2025-12-13 4237004
build: reorg app arch (#195)
