### 2025-12-13 ca1ddfa
feat: add class system keywords
