### 2025-12-13 0ef9016
feat: add class system keywords
