### 2025-12-13 8ff2693
chore(master): release 1.1.0
