### 2025-12-13 88ddbbe
feat: add parser logic for class system
