+++ src/bloch/core/lexer/token.hpp
@@ -93,6 +93,7 @@ namespace bloch::core {
         Tilde,
         Question,
         Colon,
+        Dot,
         Semicolon,
         Comma,
         Arrow,
