+++ src/bloch/core/lexer/lexer.cpp
@@ -153,6 +153,8 @@ namespace bloch::core {
                 return makeToken(TokenType::Question, "?");
             case ':':
                 return makeToken(TokenType::Colon, ":");
+            case '.':
+                return makeToken(TokenType::Dot, ".");
             case ';':
                 return makeToken(TokenType::Semicolon, ";");
             case ',':

[…]
