### 2025-12-13 20241c8
feat: add parser logic for class system
