### 2025-12-13 4e0ba8d
feat: add parser logic for class system
