### 2025-12-13 7f33c46
feat: add parser logic for class system
