### 2025-12-13 f0c21d0
feat: add parser logic for class system (#198)
