+++ src/bloch/core/ast/ast.hpp
@@ -16,6 +16,7 @@
 
 #include <memory>
 #include <string>
+#include <utility>
 #include <vector>
 
 namespace bloch::core {
@@ -43,6 +44,8 @@ namespace bloch::core {
     struct BlockStatement;
     struct AnnotationNode;
     struct Parameter;
+
+    enum class Visibility { Public, Private, Protected };
 
     // Variable Declaration
     struct VariableDeclaration : public Statement {
@@ -138,6 +141,14 @@ namespace bloch::core {
         void accept(ASTVisitor& visitor) override;
     };
 
+    // destroy expr;
+    struct DestroyStatement : public Statement {
+        std::unique_ptr<Expression> target;
+
+        DestroyStatement() = default;
+        void accept(ASTVisitor& visitor) override;
+    };
+
     // Ternary Statement
     // cond ? thenBranch : elseBranch
     struct TernaryStatement : public Statement {
@@ -220,6 +231,36 @@ namespace bloch::core {
         void accept(ASTVisitor& visitor) override;
     };
 
+    // Member access expression: object.member
+    struct MemberAccessExpression : public Expression {
+        std::unique_ptr<Expression> object;
+        std::string member;
+
+        MemberAccessExpression() = default;
+        void accept(ASTVisitor& visitor) override;
+    };
+
+    // Object creation
+    struct NewExpression : public Expression {
+        std::unique_ptr<Type> classType;
+        std::vector<std::unique_ptr<Expression>> arguments;
+
+        NewExpression() = default;
+        void accept(ASTVisitor& visitor) override;
+    };
+
+    // this
+    struct ThisExpression : public Expression {
+        ThisExpression() = default;
+        void accept(ASTVisitor& visitor) override;
+    };
+
+    // super
+    struct SuperExpression : public Expression {
+        SuperExpression() = default;
+        void accept(ASTVisitor& visitor) override;
+    };
+
     // Index Expression a[i]
     // Indexing into arrays (bounds/typing checked later).
     struct IndexExpression : public Expression {
@@ -241,7 +282,7 @@ namespace bloch::core {
     };
 
     // Parenthesized Expression
-    // (expr) — keeps source intent; may be useful for diagnostics.
+    // (expr) keeps source intent; may be useful for diagnostics.
     struct ParenthesizedExpression : public Expression {
         std::unique_ptr<Expression> expression;
 
@@ -268,6 +309,18 @@ namespace bloch::core {
         void accept(ASTVisitor& visitor) override;
     };
 
+    // object.member = value
+    struct MemberAssignmentExpression : public Expression {
+        std::unique_ptr<Expression> object;
+        std::string member;
+        std::unique_ptr<Expression> value;
+
+        MemberAssignmentExpression(std::unique_ptr<Expression> object, std::string member,
+                                   std::unique_ptr<Expression> value)
+            : object(std::move(object)), member(std::move(member)), value(std::move(value)) {}
+        void accept(ASTVisitor& visitor) override;
+    };
+
     // Array element assignment expression: collection[index] = value
     struct ArrayAssignmentExpression : public Expression {
         std::unique_ptr<Expression> collection;
@@ -289,6 +342,15 @@ namespace bloch::core {
         void accept(ASTVisitor& visitor) override;
     };
 
+    // User-defined or qualified type
+    struct NamedType : public Type {
+        std::vector<std::string> nameParts;
+
+        explicit NamedType(std::vector<std::string> parts) : nameParts(std::move(parts)) {}
[…]
