### 2025-12-13 654a851
feat: add parser logic for class system (#198)
