### 2025-12-13 a76edbb
feat: add semantic analysis for new class system
