### 2025-12-13 a83a2b8
feat: add semantic analysis for new class system
