+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -87,8 +87,7 @@ namespace bloch::core {
         return nullptr;
     }
 
-    bool SemanticAnalyser::isSubclassOf(const std::string& derived,
-                                        const std::string& base) const {
+    bool SemanticAnalyser::isSubclassOf(const std::string& derived, const std::string& base) const {
         if (derived.empty() || base.empty())
             return false;
         const ClassInfo* cur = findClass(derived);
@@ -119,9 +118,9 @@ namespace bloch::core {
             auto& name = kv.first;
             auto& m = kv.second;
             if (m.isStatic && (m.isVirtual || m.isOverride)) {
-                throw BlochError(ErrorCategory::Semantic, m.line, m.column,
-                                 "static method '" + name +
-                                     "' cannot be declared virtual or override");
+                throw BlochError(
+                    ErrorCategory::Semantic, m.line, m.column,
+                    "static method '" + name + "' cannot be declared virtual or override");
             }
             const MethodInfo* baseMethod = findMethodInHierarchy(info.base, name);
             if (m.isOverride) {
@@ -180,10 +179,9 @@ namespace bloch::core {
                     const MethodInfo* baseMethod = findMethodInHierarchy(info.base, name);
                     if (baseMethod) {
                         if (m.isStatic) {
-                            throw BlochError(
-                                ErrorCategory::Semantic, m.line, m.column,
-                                "static method '" + name +
-                                    "' cannot implement abstract base method");
+                            throw BlochError(ErrorCategory::Semantic, m.line, m.column,
+                                             "static method '" + name +
+                                                 "' cannot implement abstract base method");
                         }
                         if (baseMethod->paramTypes.size() != m.paramTypes.size() ||
                             baseMethod->returnType.className != m.returnType.className ||
@@ -218,8 +216,7 @@ namespace bloch::core {
             case Visibility::Private:
                 return owner == accessor;
             case Visibility::Protected:
-                return !accessor.empty() &&
-                       (accessor == owner || isSubclassOf(accessor, owner));
+                return !accessor.empty() && (accessor == owner || isSubclassOf(accessor, owner));
             default:
                 return false;
         }
@@ -272,14 +269,14 @@ namespace bloch::core {
                     continue;
                 if (auto field = dynamic_cast<FieldDeclaration*>(member.get())) {
                     if (info.fields.count(field->name)) {
-                        throw BlochError(ErrorCategory::Semantic, field->line, field->column,
-                                         "duplicate field '" + field->name + "' in class '" +
-                                             info.name + "'");
+                        throw BlochError(
+                            ErrorCategory::Semantic, field->line, field->column,
+                            "duplicate field '" + field->name + "' in class '" + info.name + "'");
                     }
                     if (info.isStatic && !field->isStatic) {
-                        throw BlochError(ErrorCategory::Semantic, field->line, field->column,
-                                         "static class '" + info.name +
-                                             "' cannot declare instance fields");
+                        throw BlochError(
+                            ErrorCategory::Semantic, field->line, field->column,
+                            "static class '" + info.name + "' cannot declare instance fields");
                     }
                     FieldInfo f;
                     f.visibility = field->visibility;
@@ -293,14 +290,14 @@ namespace bloch::core {
                     info.fields[field->name] = f;
                 } else if (auto method = dynamic_cast<MethodDeclaration*>(member.get())) {
                     if (info.methods.count(method->name)) {
-                        throw BlochError(ErrorCategory::Semantic, method->line, method->column,
-                                         "duplicate method '" + method->name + "' in class '" +
-                                             info.name + "'");
[…]
