### 2025-12-13 4e7fcbc
feat: add semantic analysis for class system (#200)
