+++ tests/test_semantics.cpp
@@ -415,3 +415,178 @@ TEST(SemanticTest, VoidParameterDisallowed) {
     SemanticAnalyser analyser;
     EXPECT_THROW(analyser.analyse(*program), BlochError);
 }
+
+TEST(SemanticTest, DuplicateClassMethodFails) {
+    const char* src = "class A { public function f() -> void { } public function f() -> void { } }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, PrivateFieldNotAccessibleOutsideClass) {
+    const char* src =
+        "class A { private int x = 1; public constructor() -> A { return this; } } "
+        "function main() -> void { A a = new A(); int y = a.x; }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, ProtectedFieldAccessibleInSubclass) {
+    const char* src =
+        "class Base { protected int x = 1; public constructor() -> Base { return this; } } "
+        "class Derived extends Base { public function get() -> int { return this.x; } } "
+        "function main() -> void { Derived d = new Derived(); int y = d.get(); }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_NO_THROW(analyser.analyse(*program));
+}
+
+TEST(SemanticTest, ProtectedFieldNotAccessibleOutsideHierarchy) {
+    const char* src =
+        "class Base { protected int x = 1; public constructor() -> Base { return this; } } "
+        "class Other { public function get() -> int { Base b = new Base(); return b.x; } }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, PrivateFieldNotAccessibleInSubclass) {
+    const char* src =
+        "class Base { private int x = 1; public constructor() -> Base { return this; } } "
+        "class Child extends Base { public function get() -> int { return this.x; } }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, AbstractMethodMustBeImplemented) {
+    const char* src =
+        "class A { public virtual function foo() -> void; } "
+        "class B extends A { } "
+        "function main() -> void { B b = new B(); }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, AbstractMethodImplementedAllowsInstantiation) {
+    const char* src =
+        "class A { public virtual function foo() -> void; } "
+        "class B extends A { public override function foo() -> void { return; } } "
+        "function main() -> void { B b = new B(); }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_NO_THROW(analyser.analyse(*program));
+}
+
+TEST(SemanticTest, OverrideWithoutBaseFails) {
+    const char* src = "class A { public override function foo() -> void { return; } }";
+    auto program = parseProgram(src);
+    SemanticAnalyser analyser;
+    EXPECT_THROW(analyser.analyse(*program), BlochError);
+}
+
+TEST(SemanticTest, SuperCallMustBeFirstInConstructor) {
+    const char* src =
+        "class A { public constructor() -> A { return this; } } "
[…]
