+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -14,6 +14,8 @@
 
 #include "bloch/core/semantics/semantic_analyser.hpp"
 
+#include <algorithm>
+#include <functional>
 #include <iostream>
 
 #include "bloch/core/semantics/built_ins.hpp"
@@ -23,33 +25,463 @@ namespace bloch::core {
     using support::BlochError;
     using support::ErrorCategory;
 
-    void SemanticAnalyser::analyse(Program& program) {
-        // Reset state so analyser instances are safely reusable.
-        m_symbols = SymbolTable{};
-        m_currentReturnType = ValueType::Unknown;
-        m_foundReturn = false;
-        m_functions.clear();
-        m_functionInfo.clear();
-
-        // Global scope for top-level declarations and functions.
-        ScopeGuard globalScope(*this);
-        program.accept(*this);
+    SemanticAnalyser::TypeInfo SemanticAnalyser::combine(ValueType prim, const std::string& cls) {
+        TypeInfo t;
+        t.value = prim;
+        t.className = cls;
+        return t;
     }
 
-    ValueType SemanticAnalyser::inferType(Expression* expr) const {
-        if (!expr)
-            return ValueType::Unknown;
-        if (auto lit = dynamic_cast<LiteralExpression*>(expr)) {
-            return typeFromString(lit->literalType);
+    SemanticAnalyser::TypeInfo SemanticAnalyser::typeFromAst(Type* typeNode) const {
+        if (!typeNode)
+            return combine(ValueType::Unknown, "");
+        if (auto prim = dynamic_cast<PrimitiveType*>(typeNode))
+            return combine(typeFromString(prim->name), "");
+        if (auto named = dynamic_cast<NamedType*>(typeNode)) {
+            std::string cls = named->nameParts.empty() ? "" : named->nameParts.back();
+            return combine(ValueType::Unknown, cls);
         }
+        if (dynamic_cast<VoidType*>(typeNode))
+            return combine(ValueType::Void, "");
+        if (auto arr = dynamic_cast<ArrayType*>(typeNode)) {
+            auto elem = typeFromAst(arr->elementType.get());
+            if (!elem.className.empty())
+                return combine(ValueType::Unknown, elem.className + "[]");
+            return combine(ValueType::Unknown, typeToString(elem.value) + "[]");
+        }
+        return combine(ValueType::Unknown, "");
+    }
+
+    const SemanticAnalyser::ClassInfo* SemanticAnalyser::findClass(const std::string& name) const {
+        auto it = m_classes.find(name);
+        if (it != m_classes.end())
+            return &it->second;
+        return nullptr;
+    }
+
+    SemanticAnalyser::MethodInfo* SemanticAnalyser::findMethodInHierarchy(
+        const std::string& className, const std::string& method) const {
+        const ClassInfo* cur = findClass(className);
+        while (cur) {
+            auto mit = cur->methods.find(method);
+            if (mit != cur->methods.end())
+                return const_cast<MethodInfo*>(&mit->second);
+            if (cur->base.empty())
+                break;
+            cur = findClass(cur->base);
+        }
+        return nullptr;
+    }
+
+    SemanticAnalyser::FieldInfo* SemanticAnalyser::findFieldInHierarchy(
+        const std::string& className, const std::string& field) const {
+        const ClassInfo* cur = findClass(className);
+        while (cur) {
+            auto fit = cur->fields.find(field);
+            if (fit != cur->fields.end())
+                return const_cast<FieldInfo*>(&fit->second);
+            if (cur->base.empty())
+                break;
+            cur = findClass(cur->base);
+        }
+        return nullptr;
+    }
+
+    bool SemanticAnalyser::isSubclassOf(const std::string& derived, const std::string& base) const {
[…]

[…]
