### 2025-12-13 1b85c1d
feat: add semantic analysis for class system (#200)
