### 2025-12-14 #9503 @joshiCoding
> 👋 [@kareem1925](https://github.com/kareem1925) & [@joshiCoding](https://github.com/joshiCoding), would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I will be glad to review this submission
