### 2025-12-14 #9503 @kareem1925
> 👋 [@kareem1925](https://github.com/kareem1925) & [@joshiCoding](https://github.com/joshiCoding), would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi,
I have just seen the notifications now. I will be glad to review it if I'm still needed
