### 2025-12-14 #9625 @editorialbot
[REVIEW]: Bloch: a strongly typed, hardware-agnostic, hybrid quantum programming language
**Submitting author:** @bloch-akshay (<a href="http://orcid.org/0009-0004-0465-4420">Akshay Pal</a>)
**Repository:** https://github.com/bloch-labs/bloch
**Branch with paper.md** (empty if default branch): joss-paper-submission
**Version:** v1.0.0
