### 2025-12-22 81e5480
feat: add runtime handling and gc for class system
