+++ src/bloch/core/semantics/semantic_analyser.hpp
@@ -105,6 +105,7 @@ namespace bloch::core {
             bool isStatic = false;
             bool isFinal = false;
             bool hasInitializer = false;
+            bool isTracked = false;
             TypeInfo type;
             std::string owner;
             int line = 0;
@@ -170,6 +171,7 @@ namespace bloch::core {
                                           const std::string& method) const;
         FieldInfo* findFieldInHierarchy(const std::string& className,
                                         const std::string& field) const;
+        const FieldInfo* resolveField(const std::string& name, int line, int column) const;
         bool isSubclassOf(const std::string& derived, const std::string& base) const;
         void validateOverrides(ClassInfo& info);
         void validateAbstractness(ClassInfo& info);
