+++ src/bloch/runtime/runtime_evaluator.hpp
@@ -14,8 +14,13 @@
 
 #pragma once
 
+#include <atomic>
+#include <condition_variable>
 #include <iostream>
+#include <memory>
+#include <mutex>
 #include <string>
+#include <thread>
 #include <unordered_map>
 #include <utility>
 #include <vector>
@@ -24,6 +29,10 @@
 #include "bloch/runtime/qasm_simulator.hpp"
 
 namespace bloch::runtime {
+
+    class RuntimeEvaluator;
+    struct RuntimeClass;
+    struct Object;
 
     using core::AnnotationNode;
     using core::ArrayAssignmentExpression;
@@ -34,6 +43,10 @@ namespace bloch::runtime {
     using core::BinaryExpression;
     using core::BlockStatement;
     using core::CallExpression;
+    using core::DestroyStatement;
+    using core::ConstructorDeclaration;
+    using core::DestructorDeclaration;
+    using core::FieldDeclaration;
     using core::EchoStatement;
     using core::Expression;
     using core::ExpressionStatement;
@@ -42,16 +55,23 @@ namespace bloch::runtime {
     using core::IfStatement;
     using core::IndexExpression;
     using core::LiteralExpression;
+    using core::MemberAccessExpression;
+    using core::MemberAssignmentExpression;
     using core::MeasureExpression;
     using core::MeasureStatement;
+    using core::MethodDeclaration;
+    using core::NewExpression;
     using core::Parameter;
     using core::ParenthesizedExpression;
     using core::PostfixExpression;
     using core::PrimitiveType;
     using core::Program;
+    using core::NamedType;
     using core::ResetStatement;
     using core::ReturnStatement;
     using core::Statement;
+    using core::SuperExpression;
+    using core::ThisExpression;
     using core::TernaryStatement;
     using core::Type;
     using core::UnaryExpression;
@@ -77,6 +97,9 @@ namespace bloch::runtime {
             StringArray,
             CharArray,
             QubitArray,
+            Object,
+            ObjectArray,
+            ClassRef,
             Void
         };
         Type type = Type::Void;
@@ -92,6 +115,10 @@ namespace bloch::runtime {
         std::vector<std::string> stringArray;
         std::vector<char> charArray;
         std::vector<int> qubitArray;
+        std::shared_ptr<Object> objectValue;
+        std::vector<std::shared_ptr<Object>> objectArray;
+        RuntimeClass* classRef = nullptr;
+        std::string className;
 
         Value() = default;
         explicit Value(Type t) : type(t) {}
@@ -105,17 +132,80 @@ namespace bloch::runtime {
               charValue(charVal) {}
     };
 
+    struct RuntimeTypeInfo {
+        Value::Type kind = Value::Type::Void;
+        std::string className;
+    };
+
+    struct RuntimeField {
+        std::string name;
+        RuntimeTypeInfo type;
+        bool isStatic = false;
+        bool isFinal = false;
+        bool isTracked = false;
+        Expression* initializer = nullptr;
+        bool hasInitializer = false;
+        int arraySize = -1;
+        int line = 0;
+        int column = 0;
+        size_t offset = 0;
+    };
+
+    struct RuntimeMethod;
+    struct RuntimeConstructor {
+        ConstructorDeclaration* decl = nullptr;
+        std::vector<RuntimeTypeInfo> params;
+    };
+
+    struct RuntimeMethod {
+        MethodDeclaration* decl = nullptr;
+        bool isStatic = false;
+        bool isVirtual = false;
+        bool isOverride = false;
+        size_t vtableIndex = 0;
+        RuntimeClass* owner = nullptr;
+    };
+
+    struct RuntimeClass {
+        std::string name;
+        RuntimeClass* base = nullptr;
+        bool isStatic = false;
[…]
