+++ docs/source/reference/index.md
@@ -1,0 +1,10 @@
+# Language Reference
+
+> Formal reference for the Bloch language and standard library.
+
+- Syntax and grammar (tokens, expressions, control flow).
+- Types and data structures.
+- Standard library: gates, math helpers, measurement utilities.
+- CLI/command reference.
+
+Link to grammar files or autogenerated references when available.
