+++ docs/source/examples/index.md
@@ -1,0 +1,12 @@
+# Examples
+
+The `examples/` directory contains runnable Bloch programs. Keep this page in sync with `examples/README.md` and ensure each example has a short description plus run command.
+
+| Example | Concept | Run command |
+| --- | --- | --- |
+| `01_hadamard.bloch` | Single-qubit Hadamard and measurement | `bloch examples/01_hadamard.bloch` |
+| `02_bell_state.bloch` | Bell pair creation | `bloch examples/02_bell_state.bloch` |
+| `03_deutsch_jozsa.bloch` | Deutsch–Jozsa algorithm | `bloch examples/03_deutsch_jozsa.bloch` |
+| `04_grover_search.bloch` | Grover search | `bloch examples/04_grover_search.bloch` |
+
+Add rows as new examples land. Include expected output snippets where helpful.
