+++ docs/source/contributors/index.md
@@ -1,0 +1,21 @@
+# Contributor Guide
+
+This track is for people extending Bloch itself—compiler, runtime, language, and tooling.
+
+```{toctree}
+:maxdepth: 2
+:caption: Contributors
+
+dev-setup
+workflow
+style
+release
+../architecture/index
+```
+
+## What you get here
+
+- Development environment setup and fast inner-loop tips.
+- Codebase tour and contribution workflow.
+- Coding standards, testing, and documentation expectations.
+- Release/versioning process and governance notes.
