+++ docs/source/index.md
@@ -1,0 +1,52 @@
+# Bloch Documentation
+
+```{div} release-line
+Release: **{latest_version}**
+```
+
+Bloch is a strongly typed, hardware-agnostic quantum programming language for building clear hybrid classical/quantum algorithms.
+
+```{div} card-grid
+:::{div} doc-card
+### User Guide
+Install Bloch, run quickstart circuits, and follow tutorials from Hadamard to Grover and QAOA.
+
+:::{button-link} users/index
+:color: primary
+Open User Guide →
+:::
+:::
+
+:::{div} doc-card
+### Developer Guide
+Set up the dev environment, learn the compiler/runtime architecture, and follow the contribution workflow.
+
+:::{button-link} contributors/index
+:color: primary
+Open Developer Guide →
+:::
+:::
+
+:::{div} doc-card
+### Bloch API
+Find language syntax, types, and CLI reference for day-to-day use.
+
+:::{button-link} reference/index
+:color: primary
+Open API Reference →
+:::
+:::
+```
+
+```{toctree}
+:maxdepth: 2
+:caption: Overview
+:hidden:
+
+users/index
+contributors/index
+examples/index
+reference/index
+architecture/index
+release-notes
+```
