### 2025-12-22 24d31d2
feat: added runtime support for class system
