+++ tests/test_parser.cpp
@@ -617,3 +617,46 @@ function main() -> void {
     ASSERT_NE(destroyVar, nullptr);
     EXPECT_EQ(destroyVar->name, "f");
 }
+
+TEST(ParserTest, ParsesDefaultConstructor) {
+    const char* src = R"(
+class Foo {
+    int x;
+    public constructor(int x) -> Foo = default;
+}
+)";
+    Lexer lexer(src);
+    auto tokens = lexer.tokenize();
+    Parser parser(std::move(tokens));
+    auto program = parser.parse();
+
+    ASSERT_EQ(program->classes.size(), 1u);
+    auto* foo = program->classes[0].get();
+    ASSERT_EQ(foo->members.size(), 2u);
+    auto* ctor = dynamic_cast<ConstructorDeclaration*>(foo->members[1].get());
+    ASSERT_NE(ctor, nullptr);
+    EXPECT_TRUE(ctor->isDefault);
+    EXPECT_EQ(ctor->params.size(), 1u);
+    EXPECT_EQ(ctor->params[0]->name, "x");
+    EXPECT_EQ(ctor->body, nullptr);
+}
+
+TEST(ParserTest, ParsesDefaultDestructor) {
+    const char* src = R"(
+class Foo {
+    public destructor -> void = default;
+}
+)";
+    Lexer lexer(src);
+    auto tokens = lexer.tokenize();
+    Parser parser(std::move(tokens));
+    auto program = parser.parse();
+
+    ASSERT_EQ(program->classes.size(), 1u);
+    auto* foo = program->classes[0].get();
+    ASSERT_EQ(foo->members.size(), 1u);
+    auto* dtor = dynamic_cast<DestructorDeclaration*>(foo->members[0].get());
+    ASSERT_NE(dtor, nullptr);
+    EXPECT_TRUE(dtor->isDefault);
+    EXPECT_EQ(dtor->body, nullptr);
+}
