### 2025-12-22 578400a
feat: added runtime support for class system
