+++ src/bloch/core/parser/parser.hpp
@@ -62,15 +62,12 @@ namespace bloch::core {
         // Declarations
         [[nodiscard]] std::unique_ptr<ClassMember> parseClassMember(const std::string& className,
                                                                     bool isStaticClass);
-        [[nodiscard]] std::unique_ptr<FieldDeclaration> parseFieldDeclaration(Visibility vis,
-                                                                              bool isFinal,
-                                                                              bool isStatic,
-                                                                              std::vector<std::unique_ptr<AnnotationNode>> annotations);
-        [[nodiscard]] std::unique_ptr<MethodDeclaration> parseMethodDeclaration(Visibility vis,
-                                                                                bool isStatic,
-                                                                                bool isVirtual,
-                                                                                bool isOverride,
-                                                                                std::vector<std::unique_ptr<AnnotationNode>> annotations);
+        [[nodiscard]] std::unique_ptr<FieldDeclaration> parseFieldDeclaration(
+            Visibility vis, bool isFinal, bool isStatic,
+            std::vector<std::unique_ptr<AnnotationNode>> annotations);
+        [[nodiscard]] std::unique_ptr<MethodDeclaration> parseMethodDeclaration(
+            Visibility vis, bool isStatic, bool isVirtual, bool isOverride,
+            std::vector<std::unique_ptr<AnnotationNode>> annotations);
         [[nodiscard]] std::unique_ptr<ConstructorDeclaration> parseConstructorDeclaration(
             Visibility vis, const std::string& className);
         [[nodiscard]] std::unique_ptr<DestructorDeclaration> parseDestructorDeclaration(
