### 2025-12-22 25f88a6
feat: added runtime support for class system
