+++ tests/test_runtime.cpp
@@ -482,16 +482,16 @@ TEST(RuntimeTest, ClassCtorDtorVirtualAndStatic) {
     const char* src = R"(
 class Base {
     public constructor() -> Base { echo("Base::ctor"); }
-    destructor -> void { echo("Base::dtor"); }
-    virtual function name() -> string { return "Base"; }
-    function baseOnly() -> string { return "BaseOnly"; }
+    destructor() -> void { echo("Base::dtor"); }
+    public virtual function name() -> string { return "Base"; }
+    public function baseOnly() -> string { return "BaseOnly"; }
 }
 
 class Derived extends Base {
-    static int count;
+    public static int count;
     public constructor() -> Derived { Derived.count = Derived.count + 1; echo("Derived::ctor"); }
-    destructor -> void { echo("Derived::dtor"); }
-    override function name() -> string { return "Derived"; }
+    destructor() -> void { echo("Derived::dtor"); }
+    public override function name() -> string { return "Derived"; }
 }
 
 function main() -> void {
@@ -516,7 +516,8 @@ function main() -> void {
 
 TEST(RuntimeTest, CycleCollectorReclaimsClassicalCycle) {
     const char* src =
-        "class Node { Node next; public constructor() -> Node { } } function main() -> void { Node "
+        "class Node { public Node next; public constructor() -> Node { } } function main() -> void "
+        "{ Node "
         "a = new "
         "Node(); Node b = new Node(); a.next = b; b.next = a; }";
     auto program = parseProgram(src);
@@ -529,7 +530,8 @@ TEST(RuntimeTest, CycleCollectorReclaimsClassicalCycle) {
 
 TEST(RuntimeTest, CycleCollectorSkipsTrackedCycles) {
     const char* src =
-        "class Q { @tracked qubit q; Q other; public constructor() -> Q { } } function main() -> "
+        "class Q { @tracked qubit q; public Q other; public constructor() -> Q { } } function "
+        "main() -> "
         "void { Q "
         "a = new Q(); Q b = new Q(); a.other = b; b.other = a; }";
     auto program = parseProgram(src);
