### 2025-12-22 9940fcc
feat: added runtime support for class system (#209)
