+++ src/bloch/core/ast/ast.hpp
@@ -412,8 +412,10 @@ namespace bloch::core {
         std::string name;
         std::unique_ptr<Type> fieldType;
         std::unique_ptr<Expression> initializer;
+        std::vector<std::unique_ptr<AnnotationNode>> annotations;
         bool isFinal = false;
         bool isStatic = false;
+        bool isTracked = false;
 
         FieldDeclaration() = default;
         void accept(ASTVisitor& visitor) override;
@@ -425,6 +427,8 @@ namespace bloch::core {
         std::vector<std::unique_ptr<Parameter>> params;
         std::unique_ptr<Type> returnType;
         std::unique_ptr<BlockStatement> body;
+        std::vector<std::unique_ptr<AnnotationNode>> annotations;
+        bool hasQuantumAnnotation = false;
         bool isStatic = false;
         bool isVirtual = false;
         bool isOverride = false;
@@ -437,6 +441,7 @@ namespace bloch::core {
     struct ConstructorDeclaration : public ClassMember {
         std::vector<std::unique_ptr<Parameter>> params;
         std::unique_ptr<BlockStatement> body;
+        bool isDefault = false;
 
         ConstructorDeclaration() = default;
         void accept(ASTVisitor& visitor) override;
@@ -445,6 +450,7 @@ namespace bloch::core {
     // Destructor
     struct DestructorDeclaration : public ClassMember {
         std::unique_ptr<BlockStatement> body;
+        bool isDefault = false;
 
         DestructorDeclaration() = default;
         void accept(ASTVisitor& visitor) override;
