+++ src/bloch/core/lexer/token.hpp
@@ -45,6 +45,7 @@ namespace bloch::core {
         Measure,
         Final,
         Reset,
+        Default,
 
         // Annotations
         At,
