### 2025-12-22 33a7b9c
feat: added runtime support for class system (#209)
