+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -87,6 +87,26 @@ namespace bloch::core {
         return nullptr;
     }
 
+    const SemanticAnalyser::FieldInfo* SemanticAnalyser::resolveField(const std::string& name,
+                                                                      int line, int column) const {
+        if (m_currentClass.empty())
+            return nullptr;
+        FieldInfo* field = findFieldInHierarchy(m_currentClass, name);
+        if (!field)
+            return nullptr;
+        if (!isAccessible(field->visibility, field->owner, m_currentClass)) {
+            throw BlochError(ErrorCategory::Semantic, line, column,
+                             "field '" + name + "' is not accessible here");
+        }
+        if (m_inStaticContext && !field->isStatic) {
+            throw BlochError(ErrorCategory::Semantic, line, column,
+                             "instance field '" + name +
+                                 "' cannot be referenced in a static "
+                                 "context");
+        }
+        return field;
+    }
+
     bool SemanticAnalyser::isSubclassOf(const std::string& derived, const std::string& base) const {
         if (derived.empty() || base.empty())
             return false;
@@ -283,6 +303,7 @@ namespace bloch::core {
                     f.isStatic = field->isStatic;
                     f.isFinal = field->isFinal;
                     f.hasInitializer = field->initializer != nullptr;
+                    f.isTracked = field->isTracked;
                     f.type = typeFromAst(field->fieldType.get());
                     f.owner = info.name;
                     f.line = field->line;
@@ -328,6 +349,7 @@ namespace bloch::core {
                     MethodInfo ctorInfo;
                     ctorInfo.visibility = ctor->visibility;
                     ctorInfo.hasBody = ctor->body != nullptr;
+                    ctorInfo.isDefault = ctor->isDefault;
                     ctorInfo.owner = info.name;
                     ctorInfo.line = ctor->line;
                     ctorInfo.column = ctor->column;
@@ -335,28 +357,25 @@ namespace bloch::core {
                         ctorInfo.paramTypes.push_back(typeFromAst(p->type.get()));
                     ctorInfo.returnType = combine(ValueType::Unknown, info.name);
                     info.constructors.push_back(ctorInfo);
+                    info.ctorDecls.push_back(ctor);
                 } else if (auto dtor = dynamic_cast<DestructorDeclaration*>(member.get())) {
                     if (info.isStatic) {
                         throw BlochError(
                             ErrorCategory::Semantic, dtor->line, dtor->column,
                             "static class '" + info.name + "' cannot declare destructors");
                     }
-                    if (info.hasDestructor) {
+                    if (info.hasUserDestructor) {
                         throw BlochError(
                             ErrorCategory::Semantic, dtor->line, dtor->column,
                             "class '" + info.name + "' cannot declare multiple destructors");
                     }
                     info.hasDestructor = true;
+                    info.hasUserDestructor = true;
                 }
             }
-            if (info.constructors.empty()) {
-                for (const auto& [fname, f] : info.fields) {
-                    if (f.isFinal && !f.hasInitializer) {
-                        throw BlochError(ErrorCategory::Semantic, f.line, f.column,
-                                         "final field '" + fname +
-                                             "' must be initialised or assigned in a constructor");
-                    }
-                }
+            if (!info.isStatic && info.constructors.empty()) {
+                throw BlochError(ErrorCategory::Semantic, info.line, info.column,
+                                 "class '" + info.name + "' must declare a constructor");
             }
             m_classes[info.name] = std::move(info);
         }
@@ -365,6 +384,48 @@ namespace bloch::core {
             if (!info.base.empty() && !m_classes.count(info.base)) {
                 throw BlochError(ErrorCategory::Semantic, 0, 0,
[…]
