### 2025-12-23 5e24ae7
feat: import system for class logic
