### 2025-12-23 6eabf48
chore: application direction example folder
