### 2025-12-23 e8c3df3
feat: import system for class logic
