### 2025-12-23 7acf8db
feat: import system for class logic (#210)
