+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -704,13 +704,6 @@ namespace bloch::core {
 
     void SemanticAnalyser::analyse(Program& program) {
         buildClassRegistry(program);
-        if (!program.imports.empty()) {
-            auto* first = program.imports.front().get();
-            int line = first ? first->line : 0;
-            int col = first ? first->column : 0;
-            throw BlochError(ErrorCategory::Semantic, line, col,
-                             "imports are parsed but not yet supported by the semantic analyser");
-        }
         beginScope();
         // Make class names visible as types/values (for static access).
         for (const auto& kv : m_classes) {
