### 2025-12-23 9207005
feat: import system for class logic (#210)
