+++ src/bloch/core/parser/parser.cpp
@@ -118,7 +118,7 @@ namespace bloch::core {
 
     // Main parse function
     std::unique_ptr<Program> Parser::parse() {
-        auto program = std::make_unique<Program>();
+        std::unique_ptr<Program> program = std::make_unique<Program>();
         // We alternate between function declarations and top-level statements.
         // Extra statements generated from multi-declarations are flushed as we go.
         while (!isAtEnd()) {
@@ -141,7 +141,7 @@ namespace bloch::core {
     // Top level
 
     std::unique_ptr<ImportDeclaration> Parser::parseImport() {
-        auto import = std::make_unique<ImportDeclaration>();
+        std::unique_ptr<ImportDeclaration> import = std::make_unique<ImportDeclaration>();
         const Token& importTok = previous();
         import->line = importTok.line;
         import->column = importTok.column;
@@ -173,7 +173,7 @@ namespace bloch::core {
 
         const Token& nameTok = expect(TokenType::Identifier, "Expected class name after 'class'");
 
-        auto cls = std::make_unique<ClassDeclaration>();
+        std::unique_ptr<ClassDeclaration> cls = std::make_unique<ClassDeclaration>();
         cls->name = nameTok.value;
         cls->line = nameTok.line;
         cls->column = nameTok.column;
@@ -189,7 +189,7 @@ namespace bloch::core {
 
         (void)expect(TokenType::LBrace, "Expected '{' to start class body");
         while (!check(TokenType::RBrace) && !isAtEnd()) {
-            auto member = parseClassMember(cls->name, cls->isStatic);
+            std::unique_ptr<ClassMember> member = parseClassMember(cls->name, cls->isStatic);
             if (member)
                 cls->members.push_back(std::move(member));
         }
@@ -200,7 +200,7 @@ namespace bloch::core {
 
     // Function Declaration
     std::unique_ptr<FunctionDeclaration> Parser::parseFunction() {
-        auto func = std::make_unique<FunctionDeclaration>();
+        std::unique_ptr<FunctionDeclaration> func = std::make_unique<FunctionDeclaration>();
 
         // Parse annotations
         while (check(TokenType::At)) {
@@ -232,7 +232,7 @@ namespace bloch::core {
         // Parse parameters
         (void)expect(TokenType::LParen, "Expected '(' after function name");
         while (!check(TokenType::RParen)) {
-            auto param = std::make_unique<Parameter>();
+            std::unique_ptr<Parameter> param = std::make_unique<Parameter>();
 
             param->type = parseType();
 
@@ -282,7 +282,7 @@ namespace bloch::core {
 
     std::unique_ptr<ClassMember> Parser::parseClassMember(const std::string& className,
                                                           bool isStaticClass) {
-        auto annotations = parseAnnotations();
+        std::vector<std::unique_ptr<AnnotationNode>> annotations = parseAnnotations();
 
         bool hasVisibility =
             check(TokenType::Public) || check(TokenType::Private) || check(TokenType::Protected);
@@ -320,7 +320,7 @@ namespace bloch::core {
             scanningModifiers = false;
         }
 
-        auto trailingAnnotations = parseAnnotations();
+        std::vector<std::unique_ptr<AnnotationNode>> trailingAnnotations = parseAnnotations();
         for (auto& ann : trailingAnnotations) annotations.push_back(std::move(ann));
 
         if (match(TokenType::Constructor)) {
@@ -370,7 +370,7 @@ namespace bloch::core {
     std::unique_ptr<FieldDeclaration> Parser::parseFieldDeclaration(
         Visibility vis, bool isFinal, bool isStatic,
         std::vector<std::unique_ptr<AnnotationNode>> annotations) {
[…]
