### 2025-12-26 49c17fd
fix: class constructor chain
