+++ src/bloch/runtime/runtime_evaluator.cpp
@@ -33,6 +33,8 @@ namespace bloch::runtime {
     using support::blochWarning;
     using support::ErrorCategory;
 
+    static constexpr bool kTraceConstructors = false;
+
     static std::pair<RuntimeField*, RuntimeClass*> findStaticFieldWithOwner(
         RuntimeClass* cls, const std::string& name) {
         RuntimeClass* cur = cls;
@@ -468,6 +470,12 @@ namespace bloch::runtime {
             RuntimeClass* rc = findClass(clsNode->name);
             if (!rc)
                 continue;
+            if (rc->base) {
+                // Refresh inherited layout now that base metadata is populated.
+                rc->instanceFields = rc->base->instanceFields;
+                rc->instanceFieldIndex = rc->base->instanceFieldIndex;
+                rc->vtable = rc->base->vtable;
+            }
             for (auto& member : clsNode->members) {
                 if (auto field = dynamic_cast<FieldDeclaration*>(member.get())) {
                     RuntimeField f;
@@ -757,37 +765,47 @@ namespace bloch::runtime {
         }
     }
 
+    namespace {
+        bool argumentsMatchCtor(const std::vector<Value>& args,
+                                const RuntimeConstructor& candidate) {
+            if (candidate.params.size() != args.size())
+                return false;
+            for (size_t i = 0; i < args.size(); ++i) {
+                const auto& expected = candidate.params[i];
+                const auto& actual = args[i];
+                if (expected.kind == Value::Type::Object) {
+                    if (actual.type != Value::Type::Object)
+                        return false;
+                    if (!expected.className.empty() && actual.className != expected.className)
+                        return false;
+                } else if (expected.kind == Value::Type::ObjectArray) {
+                    if (actual.type != Value::Type::ObjectArray)
+                        return false;
+                    if (!expected.className.empty() && actual.className != expected.className)
+                        return false;
+                } else if (expected.kind != actual.type) {
+                    return false;
+                }
+            }
+            return true;
+        }
+    }  // namespace
+
     void RuntimeEvaluator::runConstructorChain(RuntimeClass* cls,
                                                const std::shared_ptr<Object>& obj,
                                                ConstructorDeclaration* ctor,
                                                const std::vector<Value>& args) {
         if (!cls)
             return;
+
+        if (kTraceConstructors) {
+            std::cerr << "[ctor] " << cls->name << " args=" << args.size() << std::endl;
+        }
+
         bool savedReturn = m_hasReturn;
         m_hasReturn = false;
-        if (cls->base) {
-            ConstructorDeclaration* baseCtor = nullptr;
-            for (auto& c : cls->base->constructors) {
-                if (c.params.empty()) {
-                    baseCtor = c.decl;
-                    break;
-                }
-            }
-            runConstructorChain(cls->base, obj, baseCtor, {});
-        }
-        runFieldInitialisers(cls, obj);
-        if (ctor && ctor->isDefault) {
-            // Default constructor: bind parameters to matching fields.
-            for (size_t i = 0; i < ctor->params.size() && i < args.size(); ++i) {
-                const auto& param = ctor->params[i];
-                auto fieldMeta = findInstanceField(cls, param->name);
-                if (fieldMeta && fieldMeta->offset < obj->fields.size()) {
-                    obj->fields[fieldMeta->offset] = args[i];
-                }
-            }
-        }
-        if (!ctor)
-            return;
+
+        // Prepare context for this constructor (needed to evaluate explicit super(...) args).
         auto prevClass = m_currentClassCtx;
         bool prevStatic = m_inStaticContext;
         bool prevCtor = m_inConstructor;
[…]
