### 2025-12-26 #9625 @bloch-akshay
Hi @joshiCoding, merry Christmas! and thank you for the review feedback. To clarify, Bloch is an independent stand-alone language not a C++ wrapper. The compiler and runtime are written in C++ but a user doesn't need any knowledge of C++ or to have C++ installed to use Bloch, it is independent.

I have a lot of ideas for the language including some you mentioned above, we can catchup around those away from this thread if you are interested! Feel free to email me at akshay@bloch-labs.com 😄.

I will make changes to the docs and manuscript in the coming days, as per your guidance, thank you again for the feedback!
