### 2025-12-26 a14510b
docs: migrate to bloch repo
