+++ docs/index.md
@@ -1,0 +1,10 @@
+---
+title: Bloch Documentation
+---
+
+This folder holds the versioned Bloch docs. Pick the line that matches your toolchain:
+
+- **[Bloch v1.1.x](v1.1.x/index.md)** — current line with tighter diagnostics and preview class/import syntax.
+- **[Bloch v1.0.x](v1.0.x/index.md)** — stable functional core without class/import syntax.
+
+When publishing to the external docs site, copy this folder (including all version subfolders) into the site repo’s content path.
