### 2025-12-26 aeac502
feat: add @shots annotation
