### 2025-12-26 2cdc160
feat: add @shots annotation
