### 2025-12-26 7bb0d77
feat: add @shots annotation (#212)
