### 2025-12-26 ebb33e4
feat: add @shots annotation (#212)
