### 2025-12-29 58d3dab
docs: refresh bloch documentation
