+++ docs/v1.0.x/features.md
@@ -2,28 +2,39 @@
 title: Features
 ---
 
-Bloch v1.0.x provides the stable foundation of the language and simulator.
+Bloch 1.0.x is intentionally compact: a strongly typed, functional language that integrates classical control with quantum state evolution. Highlights below describe the execution model and the physics it exposes.
 
-## Language fundamentals
-- Types: `int`, `float`, `bit`, `string`, `char`, `qubit`, and arrays of each (fixed or dynamic). Arrays may be sized or sized at runtime (validated at runtime).
-- Entry point: `function main() -> void` is mandatory.
-- `@quantum` marks quantum helper functions; `main` is the entry point.
-- `@tracked` on `qubit`/`qubit[]` aggregates measurements at scope exit (printed when you use `--shots`).
-- Control flow: `if/else`, `for`, `while`, ternary (`cond ? stmt1 : stmt2`), arithmetic, logical, and bitwise operators.
+## Hybrid execution model
+- **Statevector simulator**: Bloch emits OpenQASM and drives an ideal statevector backend. Programs evolve pure states unless you measure or reset.
+- **Deterministic host control**: Classical branches and loops run on the host CPU while quantum operations queue in the simulator.
+- **Entry point**: `function main() -> void` is the only entry. Functions can be annotated `@quantum` to mark pure-quantum helpers; `main` itself stays classical.
 
-## Quantum programming
-- Built-in gates: `h`, `x`, `y`, `z`, `rx(qubit, theta)`, `ry(qubit, theta)`, `rz(qubit, theta)`, `cx(control, target)`.
-- Measure inline: `bit b = measure q;` or inside array literals (`{measure q0, measure q1}`).
-- Reset before reuse: `reset q;` returns a qubit to `|0⟩` and clears the measured marker.
-- Runtime enforces qubit lifetimes (warns if a qubit leaves scope unmeasured; errors on re-measuring without reset).
+## Type system
+- **Primitives**: `int`, `float` (literals require `f`), `bit`, `char`, `string`, and `qubit` handles. Arrays are fixed-size typed containers of any primitive.
+- **Immutability**: `final` enforces single assignment for variables that must stay constant, which simplifies control/flow proofs.
+- **No classes or imports**: All 1.0.x programs are functional modules in a single file.
 
-## Runtime & execution
-- Ideal statevector simulator with OpenQASM emission (`--emit-qasm`).
-- Multi-shot execution with aggregated tracked results (`--shots=N`).
-- CLI quality-of-life: `--echo=all|none` to control echo suppression; `--update` to fetch newer releases when available.
+## Quantum surface
+- **Gates**: Built-ins cover single- and two-qubit operations `h`, `x`, `y`, `z`, `rx`, `ry`, `rz`, and `cx`.
+- **Measurement and reset**: `measure q` yields a `bit`; `reset q` returns the qubit to \(|0\rangle\) and clears its measured flag.
+- **Tracked qubits**: Annotate qubit variables with `@tracked` to accumulate outcome counts across shots.
 
-## Behaviour and diagnostics
-- Parse errors for syntax issues (with line/column).
-- Semantic errors for type/lifetime violations (e.g., invalid assignment, mismatched array types, re-measuring without reset).
-- Runtime errors for division by zero, out-of-bounds, or invalid measurement targets.
-- Warnings for unmeasured qubits on scope exit (suppressed during intermediate shots when `--shots` > 1).
+Bloch models qubit evolution as:
+$$
+|\psi_{\text{out}}\rangle = U_{\text{program}} |\psi_{\text{in}}\rangle, \quad\text{with}\quad U_{\text{program}} = U_n \cdots U_2 U_1.
+$$
[…]
