+++ docs/v1.1.x/bloch-api.md
@@ -1,0 +1,13 @@
+---
+title: Bloch API
+---
+
+The 1.1.x API reference covers primitives, annotations, classes, imports, and built-ins.
+
+- [Index](reference/index.md)
+- [Primitives](reference/primitives/int.md)
+- [Functions](reference/functions.md)
+- [Annotations](reference/annotations.md)
+- [Keywords](reference/keywords.md)
+- [Built-ins](reference/builtins.md)
+- [Classes](reference/classes.md)

[…]
