### 2025-12-29 374af14
feat: add explicit casting for numeric values
