### 2025-12-29 e1e9718
feat: add explicit casting for numeric values
