### 2025-12-30 4fc715b
feat: add explicit casting for numeric values
