### 2025-12-30 dc41c8c
feat: add explicit casting for numeric values
