### 2025-12-30 a79fae9
feat: explicit casting for numeric values (#214)
