### 2025-12-30 3a199a9
feat: explicit casting for numeric values (#214)
