### 2025-12-30 6c6d4a2
feat: measure qubit arrays
