### 2025-12-30 0b93c13
feat: measure qubit arrays
