### 2025-12-30 905ee98
feat: measure qubit arrays (#215)
