### 2025-12-30 b480006
feat: measure qubit arrays (#215)
