### 2025-12-30 34c33fb
chore(master): release 1.1.0
