### 2026-01-02 ed81f8a
docs: joss review feedback
