+++ paper/paper.md
@@ -28,6 +28,10 @@ first-class annotations such as `@quantum` (to delineate quantum code) and
 `@tracked` (to stream measurements into probability tables). The language is
 implemented in C++20, ships with a comprehensive test suite, and is licensed
 under Apache-2.0 to encourage adoption in research and industrial workflows.
+By pairing static semantics with vendor-neutral QASM emission, Bloch targets
+the gap between dynamically typed Python-first SDKs (e.g., Qiskit and Cirq
+[@qiskit-zenodo;@cirq-zenodo]) and vendor-specific languages while keeping a
+single-binary, non-Python toolchain.
 
 # Statement of Need
 
@@ -38,18 +42,21 @@ simulation or hardware execution. Researchers building higher-level language
 abstractions (e.g., Silq [@silq] and Quipper [@quipper]) typically need to choose
 between tightly coupled vendor ecosystems or experimenting with new type
 systems without a reference interpreter capable of emitting standard assembly
-formats (OpenQASM 2 [@openqasm2]). Bloch fills this gap by delivering a
-self-contained toolchain that:
+formats (OpenQASM 2 [@openqasm2]). Practitioner-friendly hybrids such as
+PennyLane [@pennylane] and Microsoft Q# [@qsharp] also trade stronger typing or
+portability for deeper vendor/framework integration. Bloch fills this gap by delivering
+a self-contained toolchain that is explicitly scoped to the following objectives:
 
 * enforces a compact, explicit type system over both classical and quantum data
   so that invalid measurements, illegal `@quantum` return types, and improper
   qubit mutations are rejected at compile time;
-* produces OpenQASM output by default, which enables researchers to feed Bloch
-  programs into downstream simulators or device backends without rewriting
-  kernels; and
+* produces OpenQASM 2 output by default, which enables researchers to feed Bloch
+  programs into downstream simulators or device backends without rewriting;
 * exposes deterministic multi-shot execution with aggregated statistics so that
-  algorithm designers can validate probabilistic behaviour before running on expensive
-  quantum hardware.
+  algorithm designers can validate probabilistic behaviour before running on
+  expensive quantum hardware; and
+* stays outside heavyweight Python packaging so that experiments can be
+  reproduced from a single CLI binary rather than a coupled SDK stack.
 
 By targeting developers who are comfortable with systems languages but need a
 high-level quantum DSL, Bloch reduces the friction between research sketches
@@ -57,7 +64,7 @@ and reproducible experiments.
 
 # Design and Implementation
 
-Bloch’s architecture is intentionally modular:
+Bloch's architecture is intentionally modular:
 
 * **Front-end.** A hand-written lexer and Pratt-style parser (under
   `src/bloch/lexer` and `src/bloch/parser`) build an abstract syntax tree (AST)
@@ -80,14 +87,58 @@ Bloch’s architecture is intentionally modular:
   language constructs (e.g., an upcoming class system) without destabilising
   the default build.
 
-The project distributes platform-agnostic CMake build files and is validated on
+The following Bloch kernel produces the interaction shown in Figure 1 when run
+for a single shot:
+
+```bloch
+@quantum function main() -> void {
+    qubit q;
+    h(q);
+    bit r = measure(q);
+    reset(q);
+}
+```
+
+Figure 1 illustrates the runtime interaction between a Bloch program and the
+ideal statevector simulator for a single kernel, showing gate application,
+measurement, classical result return, and qubit reset:
+
+![Figure 1: Program-simulator message flow for a Bloch kernel](figures/bloch_execution.svg)
+
+
+The project distributes an installable binary and is validated on
[…]
