### 2026-01-05 ebf0cee
Adjust figure size for JOSS PDF rendering
