### 2026-01-05 #9625 @HectorMozo3110
Hi 👋 @bloch-akshay, thanks for the update!

I’ve gone through the paper and it looks good to me overall. I only opened a small PR ([#216](https://github.com/bloch-labs/bloch/pull/216)) to adjust Figure 1 for clarity. Please feel free to merge it if that looks fine on your end.

Regarding the other review points, I’ll work on addressing and updating them over the course of this week.

Thanks again, and great progress on the submission!
