### 2026-01-12 #9625 @kareem1925
> 👋 [@kareem1925](https://github.com/kareem1925) - I know you just started your review work a few days ago, but given that we have two complete reviews at this point, we have two options
>
> 1. You complete your review, hopefully within a couple of weeks, and we wait for this to happen
> 2. We proceed with the two existing reviews, which are sufficient for us to publish this submission in JOSS, and say that your review is no longer required
>
> If you hadn't yet started, I would take option 2, but I don't want to disvalue any existing review work that you have done, so I raise this question for you. Do you want to keep working on your review as part of this process? You could also, of course, do this informally outside of it, and open any issues in the source repo that you feel are helpful.

Thanks a lot for your mention.

I was sick and away since Christmas time. I will finish it today before 8 pm CET. Thanks for your patience, and my apologies
