+++ engineering/templates/design-template.md
@@ -1,0 +1,10 @@
+# Design: <title>
+
+## Summary
+## Goals / Non-goals
+## Architecture
+## Data model (if relevant)
+## API (if relevant)
+## Security / Observability
+## Rollout plan
+## Risks
