+++ engineering/templates/adr-template.md
@@ -1,0 +1,12 @@
+# ADR: <title>
+
+- **ID:** <id>
+- **Status:** Proposed | Accepted | Deprecated
+- **Date:** YYYY-MM-DD
+- **Owner:** <name>
+
+## Context
+## Decision
+## Alternatives
+## Consequences
+## Follow-ups
