+++ engineering/designs/index.md
@@ -1,0 +1,3 @@
+# Design Index
+
+- Add designs to `engineering/designs/` using the template in `engineering/templates/`.
