### 2026-01-18 6bf5e4b
docs: add engineering folder (#217)
