+++ engineering/adrs/index.md
@@ -1,0 +1,8 @@
+# Architecture Decision Records (ADRs)
+
+ADRs record *why* we made architectural decisions.
+
+## How to add an ADR
+Add ADRs to `engineering/adrs/` using the template in `engineering/templates/`.
+
+## ADR List
