+++ engineering/README.md
@@ -1,0 +1,16 @@
+# Engineering
+
+This folder contains **internal engineering documentation** for the Bloch compiler and runtime. 
+
+It’s intended for contributors and maintainers (not end users).
+
+## Structure
+
+- `adrs/` — Architecture Decision Records (the *why* behind major technical decisions).  
+- `designs/` — Design specs and proposals (the *how* and detailed plans that may evolve).  
+- `reference/` — Technical reference docs explaining how things work.
+- `templates/` — Reusable templates for ADRs, design docs.
+
+## Getting Started
+
+If you’re new: read `adrs/index.md`, then the relevant `reference/` docs for the subsystem you’re working on.
