### 2026-01-21 b05f59e
ci: auto-ff main in packager and clarify release flow
