### 2026-01-21 2d2a305
ci: auto-ff main in packager and clarify release flow
