### 2026-01-21 2e24b57
ci: allow hotfix in PR title
