+++ .github/workflows/pr-checks.yaml
@@ -27,6 +27,7 @@ jobs:
           types: |
             feat
             fix
+            hotfix
             perf
             refactor
             revert
