### 2026-01-23 ef4f145
fix: class constructor chaining
