### 2026-01-23 58695c5
feat: fix class constructor chaining (#222)
