+++ src/bloch/runtime/runtime_evaluator.cpp
@@ -1,4 +1,4 @@
-// Copyright 2025 Akshay Pal (https://bloch-labs.com)
+// Copyright 2026 Akshay Pal (https://bloch-labs.com)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -32,6 +32,8 @@ namespace bloch::runtime {
     using support::BlochError;
     using support::blochWarning;
     using support::ErrorCategory;
+
+    static constexpr bool kTraceConstructors = false;
 
     static std::pair<RuntimeField*, RuntimeClass*> findStaticFieldWithOwner(
         RuntimeClass* cls, const std::string& name) {
@@ -468,6 +470,12 @@ namespace bloch::runtime {
             RuntimeClass* rc = findClass(clsNode->name);
             if (!rc)
                 continue;
+            if (rc->base) {
+                // Refresh inherited layout now that base metadata is populated.
+                rc->instanceFields = rc->base->instanceFields;
+                rc->instanceFieldIndex = rc->base->instanceFieldIndex;
+                rc->vtable = rc->base->vtable;
+            }
             for (auto& member : clsNode->members) {
                 if (auto field = dynamic_cast<FieldDeclaration*>(member.get())) {
                     RuntimeField f;
@@ -731,7 +739,11 @@ namespace bloch::runtime {
                                                 const std::shared_ptr<Object>& obj) {
         if (!cls)
             return;
-        for (auto& field : cls->instanceFields) {
+        size_t startIdx = cls->base ? cls->base->instanceFields.size() : 0;
+        if (startIdx > cls->instanceFields.size())
+            startIdx = cls->instanceFields.size();
+        for (size_t i = startIdx; i < cls->instanceFields.size(); ++i) {
+            auto& field = cls->instanceFields[i];
             if (field.isStatic)
                 continue;
             auto& slot = obj->fields[field.offset];
@@ -757,37 +769,47 @@ namespace bloch::runtime {
         }
     }
 
+    namespace {
+        bool argumentsMatchCtor(const std::vector<Value>& args,
+                                const RuntimeConstructor& candidate) {
+            if (candidate.params.size() != args.size())
+                return false;
+            for (size_t i = 0; i < args.size(); ++i) {
+                const auto& expected = candidate.params[i];
+                const auto& actual = args[i];
+                if (expected.kind == Value::Type::Object) {
+                    if (actual.type != Value::Type::Object)
+                        return false;
+                    if (!expected.className.empty() && actual.className != expected.className)
+                        return false;
+                } else if (expected.kind == Value::Type::ObjectArray) {
+                    if (actual.type != Value::Type::ObjectArray)
+                        return false;
+                    if (!expected.className.empty() && actual.className != expected.className)
+                        return false;
+                } else if (expected.kind != actual.type) {
+                    return false;
+                }
+            }
+            return true;
+        }
+    }  // namespace
+
     void RuntimeEvaluator::runConstructorChain(RuntimeClass* cls,
                                                const std::shared_ptr<Object>& obj,
                                                ConstructorDeclaration* ctor,
                                                const std::vector<Value>& args) {
         if (!cls)
             return;
+
+        if (kTraceConstructors) {
+            std::cerr << "[ctor] " << cls->name << " args=" << args.size() << std::endl;
+        }
+
         bool savedReturn = m_hasReturn;
         m_hasReturn = false;
-        if (cls->base) {
-            ConstructorDeclaration* baseCtor = nullptr;
-            for (auto& c : cls->base->constructors) {
-                if (c.params.empty()) {
-                    baseCtor = c.decl;
[…]
