### 2026-01-26 b21b923
perf: only start gc thread when there are custom classes
