### 2026-01-26 20f7ab0
perf: only start gc thread when there are custom classes
