### 2026-01-26 b710654
docs: refresh --help cli message
