### 2026-01-26 5cd4dd9
docs: refresh --help cli message
