### 2026-01-26 fa364ff
docs: refresh --help cli message (#223)
