### 2026-01-26 83c3a25
docs: create in-repo documentation for users and contributors
