+++ docs/reference/README.md
@@ -1,0 +1,19 @@
+# Contributor Reference
+
+Contents:
+- `architecture.md` — end-to-end pipeline map.
+- `lexer.md` — tokenisation rules.
+- `parser.md` — parsing and AST construction.
+- `ast.md` — node inventory and visitor contract.
+- `semantics.md` — static checks and type system.
+- `module-loader.md` — import resolution and program merging.
+- `runtime-evaluator.md` — interpreter, classes, GC, tracked values.
+- `qasm-simulator.md` — statevector simulator and QASM logging.
+- `cli.md` — CLI flags, multi-shot/echo behaviour.
+- `update-manager.md` — version checks and self-update.
+- `support-and-errors.md` — diagnostics, error categories.
+- `feature-flags.md` — compile-time flags.
+- `casting.md` — casting rules and constraints.
+- `testing.md` — test layout and how to add coverage.
+- `coding-standards.md` — conventions for new code.
+- `file-index.md` — every file with first-introduced metadata.
