### 2026-01-26 a767376
docs: create in-repo documentation for users and contributors
