+++ docs/README.md
@@ -1,0 +1,9 @@
+# Bloch Documentation
+
+- `reference/` — deep technical docs for maintainers, one file per subsystem (lexer, parser, runtime, CLI, updater, etc.) plus coding standards and file index.
+
+- `user/` — language and tooling guides for Bloch users (getting started, language tour, quantum features, annotations, CLI usage, troubleshooting).
+
+Start here:
+- Contributors: read `reference/README.md`, then follow the per-component docs.
+- Users: read `user/README.md`, then the getting started and language tour.
