### 2026-01-26 edaea48
docs: create in-repo documentation for users and contributors
