+++ CONTRIBUTING.md
@@ -1,35 +1,24 @@
-# Contributing to Bloch
+# Contributing
 
-Thanks for helping shape Bloch! This guide keeps the contribution process simple while the language evolves rapidly.
+Thanks for your interest in contributing to Bloch! Please check the issues tab for open work or file a new issue if you spot a bug or have an idea.
 
----
+## Getting Started
 
-## Before You Start
-- We work from the `develop` branch; create feature branches from there and target it for pull requests. `main` is release-only and always matches the latest tag—no direct commits.
-- Please open or reference a GitHub issue so we keep discussion and tracking in one place.
-- CI enforces formatting, testing, and Conventional Commit messages, so local checks save time.
+To understand the codebase, start with the contributor docs in `docs/reference/README.md`. When you send a PR, include tests and any necessary doc updates.
 
----
-
-## Environment Setup
-
-### Prerequisites
-- C++20-capable compiler (GCC ≥ 10, Clang ≥ 10, or MSVC 19.28+)
-- CMake ≥ 3.16
-- Git
-
-### Build & Test
-
-#### Linux & macOS
+## Building from Source
+### Linux/MacOS
 ```bash
 git clone https://github.com/bloch-labs/bloch.git
 cd bloch
-scripts/build_and_test.sh
+cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
+cmake --build build --parallel
+ctest --test-dir build --output-on-failure
 ```
 
-Need debug symbols? Re-run the configure step with `-DCMAKE_BUILD_TYPE=Debug`. The `bloch` binary lives in `build/bin/`; you can execute examples directly with `./build/bin/bloch ./examples/02_bell_state.bloch`.
+Run the compiler against an example with `./build/bin/bloch examples/02_bell_state.bloch --shots=1024`.
 
-#### Windows (Visual Studio)
+### Windows (Visual Studio)
 ```powershell
 git clone https://github.com/bloch-labs/bloch.git
 cd bloch
@@ -38,70 +27,4 @@ cmake --build build --config Release --parallel
 ctest --test-dir build --config Release --output-on-failure
 ```
 
-These commands generate `build/Bloch.sln`, which you can open in Visual Studio for editing and debugging. Executables land under `build/bin/Release/` (or `build/bin/Debug/` when using `--config Debug`).
-
----
-
-## Workflow Guidelines
-
-### Branches
-Use `type/issue-number` (such as `feat/142` or `fix/275`). If you are experimenting, feel free to work on a draft branch locally and rename it before opening a PR.
-
-### Commits & PR Titles
-CI expects Conventional Commits:
-
-```
-<type>(optional-scope): <summary>
-```
-
-Examples: `feat: add runtime cache`, `fix(parser): handle trailing commas`. PR titles should mirror the final squash commit to keep history tidy. Reference related issues in the PR body if needed.
-
-### Pull Requests
-1. Push your branch and open a PR against `develop`.
-2. Fill in the PR template so reviewers know what changed and how you validated it.
-3. Run the basic checks locally (`clang-format`, build, unit tests). Our GitHub workflows rerun them plus CodeQL analysis.
-4. We use squash-merge to keep history tidy. The squash commit should match the Conventional Commit format above.
-
----
-
-## Feature Flags
-
-- Work for future features often lands behind feature flags. Review the process on the [Feature Flags wiki page](https://github.com/bloch-labs/bloch/wiki/Feature-Flags) before starting a branch.
-- When developing behind a flag, enable it during configuration (`cmake -S . -B build -DBLOCH_FEATURE_<FLAG>=ON`) and ensure the runtime helpers report the expected state.
-- Update the documentation entries whenever you introduce or retire a flag.
-
----
-
[…]
