### 2026-01-26 6ff283d
docs: create in-repo documentation for users and contributors
