+++ README.md
@@ -1,109 +1,37 @@
-# Bloch - A Strongly Typed, Hybrid Quantum Programming
+# Bloch
 
-Bloch is an open-source, strongly typed hybrid quantum programming language. The language is designed to be hardware-agnostic, so you can move from research sketches to production-grade circuits without switching stacks.
+Bloch is an open-source, strongly typed hybrid quantum programming language. It’s designed for people who want classical and quantum code side by side. The toolchain is hardware-agnostic so you can move between simulators and future backends.
 
-## Highlights
-- Seamless classical/quantum integration with clear type boundaries.
-- Readable, expressive syntax that stays approachable under load.
-- Hardware-agnostic execution across simulators and future hardware backends.
-- Write once, run anywhere as targets evolve.
-
-## See It In Action
-![Demo](demo/demo.svg)
-Prepare an entangled Bell state, then measure both qubits 8192 times using a tracked array to observe the expected 00 / 11 correlation.
-
-```bloch
-@quantum
-function createBellState(qubit a, qubit b) -> void {
-    h(a);
-    cx(a, b);
-}
-
-function main() -> void {
-    @tracked qubit[2] qreg;
-
-    createBellState(qreg[0], qreg[1]);
-
-    bit[2] out = {measure qreg[0], measure qreg[1]};
-
-    echo(out);
-}
+## 🚀 Install 
+- Linux/macOS:
 ```
-
-## Get Started
-Bloch is available as source plus manually produced release archives until the automated pipeline lands. The instructions below cover installing pre-built artifacts as well as building from source.
-
-### Install a release build
-
-#### Linux & macOS
-```bash
 curl -fsSL https://raw.githubusercontent.com/bloch-labs/bloch/HEAD/scripts/install.sh | bash -s -- latest
 ```
-
-Swap `latest` for a specific tag (eg `v1.0.0`). The script downloads the matching archive for your platform, verifies its checksum, installs the `bloch` binary into a writable directory, and updates your shell profile if needed. Add `INSTALL_DIR=/custom/path` before `bash` to override the destination.
-You can also self-update later with `bloch --update`; major version jumps will prompt for confirmation and link the changelog.
-
-#### Windows
-```powershell
+- Windows: 
+```
 powershell -NoProfile -ExecutionPolicy Bypass -Command "iwr https://raw.githubusercontent.com/bloch-labs/bloch/HEAD/scripts/install.ps1 -UseBasicParsing -OutFile $env:TEMP\\bloch-install.ps1; & $env:TEMP\\bloch-install.ps1 -Version latest"
 ```
+- From source: See `CONTRIBUTING.md`
 
-Swap `-Version latest` for any tag you want. The script downloads the Windows ZIP, verifies its checksum, installs `bloch.exe` into `%LOCALAPPDATA%\Programs\Bloch` by default, and adds that directory to your user `PATH` (opens in new shells).
-You can self-update later with `bloch --update`; major version jumps prompt for confirmation and show the changelog link.
+## 💻 VS Code Extension 
+Syntax highlighting: https://github.com/bloch-labs/bloch-vscode
 
-### Build from source
+## 📚 Documentation 
+- User guides: `docs/user/README.md`
+- Contributor docs: `docs/reference/README.md`
 
-#### Linux & macOS
-```bash
-git clone https://github.com/bloch-labs/bloch.git
-cd bloch
-cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
-cmake --build build --parallel
-ctest --test-dir build --output-on-failure
-```
+## 📝 Cite 
[…]
