### 2026-01-26 e97cc7e
docs: create in-repo documentation for users and contributors
