### 2026-01-26 cabe081
feat: add a bloch stdlib
