### 2026-01-26 373f160
feat: add null primitive
