+++ src/bloch/runtime/runtime_evaluator.cpp
@@ -47,6 +47,10 @@ namespace bloch::runtime {
         return {nullptr, nullptr};
     }
 
+    static bool isNullReference(const Value& v) {
+        return v.type == Value::Type::Object && !v.objectValue;
+    }
+
     static std::string valueToString(const Value& v) {
         // Pretty-print a runtime value for echo and tracked summaries.
         std::ostringstream oss;
@@ -117,6 +121,8 @@ namespace bloch::runtime {
                 oss << "}";
                 return oss.str();
             case Value::Type::Object:
+                if (isNullReference(v))
+                    return "null";
                 if (v.objectValue && v.objectValue->cls)
                     return "<" + v.objectValue->cls->name + " object>";
                 return "<object>";
@@ -129,7 +135,7 @@ namespace bloch::runtime {
                     if (obj && obj->cls)
                         oss << "<" << obj->cls->name << ">";
                     else
-                        oss << "<object>";
+                        oss << "null";
                 }
                 oss << "}";
                 return oss.str();
@@ -356,7 +362,8 @@ namespace bloch::runtime {
             if (fit != it->end()) {
                 Value newVal = v;
                 if (fit->second.value.type == Value::Type::Object &&
-                    newVal.type == Value::Type::Object && !fit->second.value.className.empty()) {
+                    newVal.type == Value::Type::Object && newVal.objectValue &&
+                    !fit->second.value.className.empty()) {
                     newVal.className = fit->second.value.className;
                 }
                 fit->second.value = newVal;
@@ -371,20 +378,21 @@ namespace bloch::runtime {
                 if (field && field->offset < thisObj->fields.size()) {
                     Value newVal = v;
                     const Value& existing = thisObj->fields[field->offset];
-                    if (existing.type == Value::Type::Object &&
-                        newVal.type == Value::Type::Object && !existing.className.empty()) {
-                        newVal.className = existing.className;
-                    }
-                    thisObj->fields[field->offset] = newVal;
-                    return;
+                if (existing.type == Value::Type::Object &&
+                    newVal.type == Value::Type::Object && newVal.objectValue &&
+                    !existing.className.empty()) {
+                    newVal.className = existing.className;
                 }
+                thisObj->fields[field->offset] = newVal;
+                return;
+            }
             }
             auto [field, owner] = findStaticFieldWithOwner(m_currentClassCtx, name);
             if (field && owner && field->offset < owner->staticStorage.size()) {
                 Value newVal = v;
                 const Value& existing = owner->staticStorage[field->offset];
                 if (existing.type == Value::Type::Object && newVal.type == Value::Type::Object &&
-                    !existing.className.empty()) {
+                    newVal.objectValue && !existing.className.empty()) {
                     newVal.className = existing.className;
                 }
                 owner->staticStorage[field->offset] = newVal;
@@ -785,6 +793,8 @@ namespace bloch::runtime {
                 const auto& expected = candidate.params[i];
                 const auto& actual = args[i];
                 if (expected.kind == Value::Type::Object) {
+                    if (isNullReference(actual))
+                        continue;
                     if (actual.type != Value::Type::Object)
                         return false;
                     if (!expected.className.empty() && actual.className != expected.className)
@@ -1056,10 +1066,8 @@ namespace bloch::runtime {
                             v.qubitArray[i] = allocateTrackedQubit(var->name);
                     }
                 }
-            } else if (auto named = dynamic_cast<NamedType*>(var->varType.get())) {
[…]
