### 2026-01-26 0399297
feat: add null primitive
