### 2026-01-26 1c2a9e1
feat: add null primitive
