+++ src/bloch/runtime/runtime_evaluator.hpp
@@ -61,9 +61,9 @@ namespace bloch::runtime {
     using core::MemberAccessExpression;
     using core::MemberAssignmentExpression;
     using core::MethodDeclaration;
-    using core::NullLiteralExpression;
     using core::NamedType;
     using core::NewExpression;
+    using core::NullLiteralExpression;
     using core::Parameter;
     using core::ParenthesizedExpression;
     using core::PostfixExpression;
