### 2026-01-26 538ae3a
feat: add null primitive (#225)
