+++ src/bloch/runtime/runtime_evaluator.hpp
@@ -63,6 +63,7 @@ namespace bloch::runtime {
     using core::MethodDeclaration;
     using core::NamedType;
     using core::NewExpression;
+    using core::NullLiteralExpression;
     using core::Parameter;
     using core::ParenthesizedExpression;
     using core::PostfixExpression;
