+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -528,6 +528,8 @@ namespace bloch::core {
     SemanticAnalyser::TypeInfo SemanticAnalyser::inferTypeInfo(Expression* expr) const {
         if (!expr)
             return combine(ValueType::Unknown, "");
+        if (dynamic_cast<NullLiteralExpression*>(expr))
+            return combine(ValueType::Null, "");
         if (auto lit = dynamic_cast<LiteralExpression*>(expr))
             return combine(typeFromString(lit->literalType), "");
         if (auto var = dynamic_cast<VariableExpression*>(expr)) {
@@ -794,6 +796,7 @@ namespace bloch::core {
             }
         }
         if (node.initializer) {
+            TypeInfo initInfo = inferTypeInfo(node.initializer.get());
             // Disallow initialisation of qubit arrays
             if (auto arr = dynamic_cast<ArrayType*>(node.varType.get())) {
                 if (auto elem = dynamic_cast<PrimitiveType*>(arr->elementType.get())) {
@@ -801,6 +804,10 @@ namespace bloch::core {
                         throw BlochError(ErrorCategory::Semantic, node.line, node.column,
                                          "qubit[] cannot be initialised");
                     }
+                }
+                if (initInfo.value == ValueType::Null) {
+                    throw BlochError(ErrorCategory::Semantic, node.line, node.column,
+                                     "initializer for '" + node.name + "' cannot be null");
                 }
             }
             if (auto call = dynamic_cast<CallExpression*>(node.initializer.get())) {
@@ -814,7 +821,7 @@ namespace bloch::core {
             // Validate initializer expression first so cast errors surface before type checks.
             node.initializer->accept(*this);
             if (auto primType = tinfo.value; primType != ValueType::Unknown) {
-                ValueType initT = inferTypeInfo(node.initializer.get()).value;
+                ValueType initT = initInfo.value;
                 if (initT != ValueType::Unknown && initT != primType) {
                     if (primType == ValueType::Bit) {
                         if (auto lit = dynamic_cast<LiteralExpression*>(node.initializer.get())) {
@@ -837,7 +844,7 @@ namespace bloch::core {
                                          typeToString(initT) + "'");
                 }
             } else if (!tinfo.className.empty()) {
-                auto initT = inferTypeInfo(node.initializer.get());
+                auto initT = initInfo;
                 if (!initT.className.empty() && initT.className != tinfo.className) {
                     throw BlochError(
                         ErrorCategory::Semantic, node.line, node.column,
@@ -958,7 +965,7 @@ namespace bloch::core {
     void SemanticAnalyser::visit(DestroyStatement& node) {
         if (node.target) {
             auto t = inferTypeInfo(node.target.get());
-            if (t.className.empty()) {
+            if (t.className.empty() && t.value != ValueType::Null) {
                 throw BlochError(ErrorCategory::Semantic, node.line, node.column,
                                  "'destroy' requires a class reference");
             }
@@ -972,8 +979,21 @@ namespace bloch::core {
                 throw BlochError(ErrorCategory::Semantic, node.line, node.column,
                                  "Cannot assign to final variable '" + node.name + "'");
             }
-            if (node.value)
+            if (node.value) {
+                auto valType = inferTypeInfo(node.value.get());
+                if (valType.value == ValueType::Null) {
+                    TypeInfo target = getVariableType(node.name);
+                    bool targetIsArray =
+                        target.className.size() >= 2 &&
+                        target.className.rfind("[]") == target.className.size() - 2;
+                    bool targetIsClass = !target.className.empty() && !targetIsArray;
+                    if (!targetIsClass) {
+                        throw BlochError(ErrorCategory::Semantic, node.line, node.column,
[…]
