### 2026-01-26 94f8646
feat: add null primitive (#225)
