### 2026-01-26 5b1d7cb
feat: implement method overloading
