+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -1,4 +1,4 @@
-// Copyright 2025 Akshay Pal (https://bloch-labs.com)
+// Copyright 2026 Akshay Pal (https://bloch-labs.com)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -32,6 +32,66 @@ namespace bloch::core {
         return t;
     }
 
+    namespace {
+        std::string methodSignatureLabel(const std::string& name,
+                                         const std::vector<SemanticAnalyser::TypeInfo>& params) {
+            std::ostringstream oss;
+            oss << name << "(";
+            for (size_t i = 0; i < params.size(); ++i) {
+                if (i)
+                    oss << ",";
+                if (!params[i].className.empty())
+                    oss << params[i].className;
+                else
+                    oss << typeToString(params[i].value);
+            }
+            oss << ")";
+            return oss.str();
+        }
+
+        bool paramTypesEqual(const std::vector<SemanticAnalyser::TypeInfo>& a,
+                             const std::vector<SemanticAnalyser::TypeInfo>& b) {
+            if (a.size() != b.size())
+                return false;
+            for (size_t i = 0; i < a.size(); ++i) {
+                if (a[i].className != b[i].className || a[i].value != b[i].value)
+                    return false;
+            }
+            return true;
+        }
+
+        bool paramsMatchCall(const std::vector<SemanticAnalyser::TypeInfo>& expected,
+                             const std::vector<SemanticAnalyser::TypeInfo>& actual) {
+            if (expected.size() != actual.size())
+                return false;
+            for (size_t i = 0; i < expected.size(); ++i) {
+                const auto& exp = expected[i];
+                const auto& act = actual[i];
+                bool actIsArray =
+                    act.className.size() >= 2 &&
+                    act.className.rfind("[]") == act.className.size() - 2;
+                if (!exp.className.empty()) {
+                    if (act.value == ValueType::Null)
+                        continue;  // nullable class
+                    if (act.value != ValueType::Unknown && act.className.empty())
+                        return false;
+                    if (!act.className.empty() && act.className != exp.className)
+                        return false;
+                } else if (exp.value != ValueType::Unknown) {
+                    if (act.value == ValueType::Unknown)
+                        continue;
+                    if (act.value == ValueType::Null)
+                        return false;
+                    if (act.value != exp.value)
+                        return false;
+                } else if (!exp.className.empty() && actIsArray) {
+                    return false;
+                }
+            }
+            return true;
+        }
+    }  // namespace
+
     SemanticAnalyser::TypeInfo SemanticAnalyser::typeFromAst(Type* typeNode) const {
         if (!typeNode)
             return combine(ValueType::Unknown, "");
@@ -60,17 +120,30 @@ namespace bloch::core {
     }
 
     SemanticAnalyser::MethodInfo* SemanticAnalyser::findMethodInHierarchy(
-        const std::string& className, const std::string& method) const {
+        const std::string& className, const std::string& method,
+        const std::vector<TypeInfo>* params) const {
         const ClassInfo* cur = findClass(className);
+        std::vector<MethodInfo*> matches;
         while (cur) {
             auto mit = cur->methods.find(method);
-            if (mit != cur->methods.end())
-                return const_cast<MethodInfo*>(&mit->second);
+            if (mit != cur->methods.end()) {
+                if (!params) {
+                    return const_cast<MethodInfo*>(&mit->second.front());
+                }
+                for (auto& cand : mit->second) {
+                    if (paramsMatchCall(cand.paramTypes, *params))
+                        matches.push_back(const_cast<MethodInfo*>(&cand));
+                }
+                if (!matches.empty())
[…]

[…]
