### 2026-01-26 bf999b3
feat: implement method overloading
