### 2026-01-26 ec05a7d
feat: implement method overloading (#226)
