### 2026-01-26 d298654
feat: implement method overloading (#226)
