### 2026-01-28 #9625 @danielskatz
@bloch-akshay - I don't see any issues with your paper, so at this point, could you:
