### 2026-01-28 8858ada
feat: implement generics
