### 2026-01-28 de10b7c
chore: fix manual packager tag reachability + add checksums.txt (#228)
