+++ .github/workflows/manual-packager.yaml
@@ -88,12 +88,15 @@ jobs:
         uses: actions/checkout@v4
         with:
           ref: refs/tags/${{ inputs.tag }}
+          fetch-depth: 0
+          fetch-tags: true
 
       - name: Verify tag is on main
         shell: bash
         run: |
-          git fetch --depth=1 origin main
-          if ! git merge-base --is-ancestor "${{ inputs.tag }}" origin/main; then
+          set -euo pipefail
+          git fetch origin main --tags --force --prune
+          if ! git merge-base --is-ancestor "refs/tags/${{ inputs.tag }}" origin/main; then
             echo "Tag ${{ inputs.tag }} is not reachable from origin/main." >&2
             exit 1
           fi
@@ -242,6 +245,13 @@ jobs:
         with:
           pattern: bloch-${{ inputs.tag }}-*
           path: dist
+
+      - name: Build global checksums.txt
+        shell: bash
+        run: |
+          set -euo pipefail
+          find dist -name 'checksums-*.txt' -type f -print -exec cat {} \; > dist/checksums.txt
+          ls -l dist/checksums.txt
 
       - name: Create GitHub release
         uses: softprops/action-gh-release@v1
