### 2026-01-28 #9625 @ghost
Hi @danielskatz thank you very much! I have generated and double checked the Zenodo archive, the DOI is https://doi.org/10.5281/zenodo.18407424.

Would like to extend a huge thanks to all the reviewers @joshiCoding, @HectorMozo3110 and @kareem1925 for their work!
