### 2026-01-28 fc6b78d
feat: implement generics
