### 2026-01-28 f30cc94
feat: implement generics
