+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -192,7 +192,8 @@ namespace bloch::core {
         }
         TypeInfo out = t;
         out.typeArgs.clear();
-        for (const auto& a : t.typeArgs) out.typeArgs.push_back(substituteTypeParams(a, params, args));
+        for (const auto& a : t.typeArgs)
+            out.typeArgs.push_back(substituteTypeParams(a, params, args));
         auto isArray = [](const std::string& name) {
             return name.size() >= 2 && name.rfind("[]") == name.size() - 2;
         };
@@ -211,8 +212,7 @@ namespace bloch::core {
         return res;
     }
 
-    void SemanticAnalyser::validateTypeApplication(const TypeInfo& t, int line,
-                                                   int column) const {
+    void SemanticAnalyser::validateTypeApplication(const TypeInfo& t, int line, int column) const {
         if (t.className.empty())
             return;
         const ClassInfo* info = findClass(t.className);
@@ -234,10 +234,11 @@ namespace bloch::core {
                                      "type argument '" + typeLabel(actual) +
                                          "' does not satisfy bound '" + typeLabel(bound) + "'");
                 }
-                if (actual.className != bound.className && !isSubclassOf(actual.className, bound.className)) {
+                if (actual.className != bound.className &&
+                    !isSubclassOf(actual.className, bound.className)) {
                     throw BlochError(ErrorCategory::Semantic, line, column,
-                                     "type argument '" + typeLabel(actual) + "' does not satisfy bound '" +
-                                         typeLabel(bound) + "'");
+                                     "type argument '" + typeLabel(actual) +
+                                         "' does not satisfy bound '" + typeLabel(bound) + "'");
                 }
             }
         }
@@ -369,9 +370,8 @@ namespace bloch::core {
                         ErrorCategory::Semantic, m.line, m.column,
                         "static method '" + m.name + "' cannot be declared virtual or override");
                 }
-                const MethodInfo* baseMethod =
-                    findMethodInHierarchy(combine(ValueType::Unknown, info.base), m.name,
-                                          &m.paramTypes);
+                const MethodInfo* baseMethod = findMethodInHierarchy(
+                    combine(ValueType::Unknown, info.base), m.name, &m.paramTypes);
                 if (m.isOverride) {
                     if (!baseMethod) {
                         throw BlochError(
@@ -419,9 +419,8 @@ namespace bloch::core {
                 if (m.hasBody) {
                     auto it = std::find(required.begin(), required.end(), m.signature);
                     if (it != required.end()) {
-                        const MethodInfo* baseMethod =
-                            findMethodInHierarchy(combine(ValueType::Unknown, info.base), m.name,
-                                                  &m.paramTypes);
+                        const MethodInfo* baseMethod = findMethodInHierarchy(
+                            combine(ValueType::Unknown, info.base), m.name, &m.paramTypes);
                         if (baseMethod) {
                             if (m.isStatic) {
                                 throw BlochError(ErrorCategory::Semantic, m.line, m.column,
@@ -821,8 +820,8 @@ namespace bloch::core {
                 if (bi != builtInGates.end())
                     return combine(bi->second.returnType, "");
                 if (!m_currentClass.empty()) {
-                    auto* method = findMethodInHierarchy(combine(ValueType::Unknown, m_currentClass),
-                                                         callee->name, &argTypes);
+                    auto* method = findMethodInHierarchy(
+                        combine(ValueType::Unknown, m_currentClass), callee->name, &argTypes);
                     if (method) {
                         if (!method->isStatic && m_inStaticContext)
                             return combine(ValueType::Unknown, "");
@@ -850,8 +849,8 @@ namespace bloch::core {
                                     binding[expected.className] = actual;
                                     return;
[…]
