### 2026-01-28 7e7cd2f
feat: implement generics (#230)
