+++ src/bloch/core/semantics/semantic_analyser.cpp
@@ -17,6 +17,7 @@
 #include <algorithm>
 #include <functional>
 #include <iostream>
+#include <unordered_map>
 
 #include "bloch/core/semantics/built_ins.hpp"
 
@@ -32,6 +33,47 @@ namespace bloch::core {
         return t;
     }
 
+    bool SemanticAnalyser::typeEquals(const TypeInfo& a, const TypeInfo& b) {
+        if (a.isTypeParam || b.isTypeParam)
+            return a.isTypeParam && b.isTypeParam && a.className == b.className;
+        if (a.value != b.value)
+            return false;
+        if (a.className != b.className)
+            return false;
+        if (a.typeArgs.size() != b.typeArgs.size())
+            return false;
+        for (size_t i = 0; i < a.typeArgs.size(); ++i) {
+            if (!typeEquals(a.typeArgs[i], b.typeArgs[i]))
+                return false;
+        }
+        return true;
+    }
+
+    std::string SemanticAnalyser::typeLabel(const TypeInfo& t) {
+        if (t.isTypeParam)
+            return t.className;
+        auto isArray = [](const std::string& name) {
+            return name.size() >= 2 && name.rfind("[]") == name.size() - 2;
+        };
+        if (isArray(t.className) && !t.typeArgs.empty()) {
+            return typeLabel(t.typeArgs.front()) + "[]";
+        }
+        if (!t.className.empty()) {
+            std::string res = t.className;
+            if (!t.typeArgs.empty()) {
+                res += "<";
+                for (size_t i = 0; i < t.typeArgs.size(); ++i) {
+                    if (i)
+                        res += ",";
+                    res += typeLabel(t.typeArgs[i]);
+                }
+                res += ">";
+            }
+            return res;
+        }
+        return typeToString(t.value);
+    }
+
     namespace {
         std::string methodSignatureLabel(const std::string& name,
                                          const std::vector<SemanticAnalyser::TypeInfo>& params) {
@@ -40,10 +82,7 @@ namespace bloch::core {
             for (size_t i = 0; i < params.size(); ++i) {
                 if (i)
                     oss << ",";
-                if (!params[i].className.empty())
-                    oss << params[i].className;
-                else
-                    oss << typeToString(params[i].value);
+                oss << SemanticAnalyser::typeLabel(params[i]);
             }
             oss << ")";
             return oss.str();
@@ -54,7 +93,7 @@ namespace bloch::core {
             if (a.size() != b.size())
                 return false;
             for (size_t i = 0; i < a.size(); ++i) {
-                if (a[i].className != b[i].className || a[i].value != b[i].value)
+                if (!SemanticAnalyser::typeEquals(a[i], b[i]))
                     return false;
             }
             return true;
@@ -67,14 +106,25 @@ namespace bloch::core {
             for (size_t i = 0; i < expected.size(); ++i) {
                 const auto& exp = expected[i];
                 const auto& act = actual[i];
-                bool actIsArray = act.className.size() >= 2 &&
-                                  act.className.rfind("[]") == act.className.size() - 2;
+                bool expIsArray = exp.className.size() >= 2 &&
+                                  exp.className.rfind("[]") == exp.className.size() - 2;
                 if (!exp.className.empty()) {
-                    if (act.value == ValueType::Null)
+                    if (exp.isTypeParam) {
+                        if (act.value != ValueType::Unknown && act.className.empty())
+                            return false;  // primitives cannot satisfy a type parameter
+                        continue;
+                    }
+                    if (act.value == ValueType::Null) {
+                        if (expIsArray)
+                            return false;
                         continue;  // nullable class
-                    if (act.value != ValueType::Unknown && act.className.empty())
+                    }
[…]
