+++ src/bloch/core/semantics/semantic_analyser.hpp
@@ -36,8 +36,13 @@ namespace bloch::core {
         struct TypeInfo {
             ValueType value = ValueType::Unknown;
             std::string className;  // non-empty when referring to a user-defined class
+            std::vector<TypeInfo> typeArgs;
+            bool isTypeParam = false;
             bool isClass() const { return !className.empty(); }
         };
+
+        static bool typeEquals(const TypeInfo& a, const TypeInfo& b);
+        static std::string typeLabel(const TypeInfo& t);
 
         // Visitors
         void visit(VariableDeclaration& node) override;
@@ -80,6 +85,7 @@ namespace bloch::core {
         void visit(VoidType& node) override;
 
         void visit(Parameter& node) override;
+        void visit(TypeParameter& node) override;
         void visit(AnnotationNode& node) override;
         void visit(ImportDeclaration& node) override;
         void visit(FieldDeclaration& node) override;
@@ -137,6 +143,11 @@ namespace bloch::core {
             bool isAbstract = false;
             bool hasDestructor = true;       // implicit default exists
             bool hasUserDestructor = false;  // true if explicitly declared
+            struct TypeParamInfo {
+                std::string name;
+                TypeInfo bound;
+            };
+            std::vector<TypeParamInfo> typeParams;
             std::vector<std::string> abstractMethods;
             std::unordered_map<std::string, FieldInfo> fields;
             std::unordered_map<std::string, std::vector<MethodInfo>> methods;
@@ -155,6 +166,9 @@ namespace bloch::core {
         bool m_allowSuperConstructorCall = false;
         std::string m_currentMethod;
         bool m_currentMethodIsOverride = false;
+
+        // Track full type information (including generic arguments) alongside the symbol table.
+        std::vector<std::unordered_map<std::string, TypeInfo>> m_typeStack;
 
         // Helpers
         void beginScope();
@@ -175,10 +189,9 @@ namespace bloch::core {
         TypeInfo inferTypeInfo(Expression* expr) const;
         void buildClassRegistry(Program& program);
         const ClassInfo* findClass(const std::string& name) const;
-        MethodInfo* findMethodInHierarchy(const std::string& className, const std::string& method,
+        MethodInfo* findMethodInHierarchy(const TypeInfo& classType, const std::string& method,
                                           const std::vector<TypeInfo>* params = nullptr) const;
-        FieldInfo* findFieldInHierarchy(const std::string& className,
-                                        const std::string& field) const;
+        FieldInfo* findFieldInHierarchy(const TypeInfo& classType, const std::string& field) const;
         const FieldInfo* resolveField(const std::string& name, int line, int column) const;
         bool isSubclassOf(const std::string& derived, const std::string& base) const;
         void validateOverrides(ClassInfo& info);
@@ -191,6 +204,18 @@ namespace bloch::core {
 
         // Type inference
         std::optional<int> evaluateConstInt(Expression* expr) const;
+
+        // Generics helpers
+        std::vector<ClassInfo::TypeParamInfo> m_currentTypeParams;
+        bool m_inClassRegistryBuild = false;
+        TypeInfo substituteTypeParams(const TypeInfo& t,
+                                      const std::vector<ClassInfo::TypeParamInfo>& params,
+                                      const std::vector<TypeInfo>& args) const;
+        std::vector<TypeInfo> substituteMany(const std::vector<TypeInfo>& types,
+                                             const std::vector<ClassInfo::TypeParamInfo>& params,
+                                             const std::vector<TypeInfo>& args) const;
+        void validateTypeApplication(const TypeInfo& t, int line, int column) const;
+        std::optional<TypeInfo> getTypeParamBound(const std::string& name) const;
     };
 
 }  // namespace bloch::core
