### 2026-01-28 dec471f
feat: implement generics (#230)
