+++ src/bloch/runtime/runtime_evaluator.cpp
@@ -21,6 +21,7 @@
 #include <iomanip>
 #include <optional>
 #include <sstream>
+#include <unordered_map>
 #include <utility>
 
 #include "bloch/core/semantics/built_ins.hpp"
@@ -32,6 +33,12 @@ namespace bloch::runtime {
     using support::BlochError;
     using support::blochWarning;
     using support::ErrorCategory;
+
+    namespace {
+        std::string typeKey(const RuntimeTypeInfo& t);
+        std::unique_ptr<Type> runtimeTypeToAst(const RuntimeTypeInfo& rt);
+        RuntimeTypeInfo typeInfoFromValue(const Value& v);
+    }
 
     static constexpr bool kTraceConstructors = false;
 
@@ -146,7 +153,8 @@ namespace bloch::runtime {
         }
     }
 
-    RuntimeTypeInfo RuntimeEvaluator::typeInfoFromAst(Type* type) const {
+    RuntimeTypeInfo RuntimeEvaluator::typeInfoFromAst(
+        Type* type, const std::unordered_map<std::string, RuntimeTypeInfo>& subst) const {
         RuntimeTypeInfo info;
         if (!type)
             return info;
@@ -164,11 +172,23 @@ namespace bloch::runtime {
             else if (prim->name == "qubit")
                 info.kind = Value::Type::Qubit;
         } else if (auto named = dynamic_cast<NamedType*>(type)) {
-            info.kind = Value::Type::Object;
-            if (!named->nameParts.empty())
-                info.className = named->nameParts.back();
+            // Type parameter substitution
+            if (!named->nameParts.empty()) {
+                std::string base = named->nameParts.back();
+                auto it = subst.find(base);
+                if (it != subst.end())
+                    return it->second;
+                info.kind = Value::Type::Object;
+                info.className = base;
+                for (auto& arg : named->typeArguments)
+                    info.typeArgs.push_back(typeInfoFromAst(arg.get(), subst));
+                info.className = typeKey(info);
+                // Strip to base if no args (non-generic class)
+                if (info.typeArgs.empty())
+                    info.className = base;
+            }
         } else if (auto arr = dynamic_cast<ArrayType*>(type)) {
-            auto elem = typeInfoFromAst(arr->elementType.get());
+            auto elem = typeInfoFromAst(arr->elementType.get(), subst);
             switch (elem.kind) {
                 case Value::Type::Int:
                     info.kind = Value::Type::IntArray;
@@ -188,17 +208,165 @@ namespace bloch::runtime {
                 case Value::Type::Qubit:
                     info.kind = Value::Type::QubitArray;
                     break;
-                case Value::Type::Object:
+                default:
                     info.kind = Value::Type::ObjectArray;
                     info.className = elem.className;
-                    break;
-                default:
+                    info.typeArgs.push_back(elem);
                     break;
             }
         } else if (dynamic_cast<VoidType*>(type)) {
             info.kind = Value::Type::Void;
         }
         return info;
+    }
+
+    namespace {
+        std::string typeKey(const RuntimeTypeInfo& t) {
+            if (!t.className.empty()) {
+                if (t.typeArgs.empty())
+                    return t.className;
+                std::ostringstream oss;
+                oss << t.className;
+                // If className already includes args (e.g. from nesting), avoid duplication.
+                if (t.className.find('<') == std::string::npos) {
+                    oss << "<";
+                    for (size_t i = 0; i < t.typeArgs.size(); ++i) {
+                        if (i)
+                            oss << ",";
+                        oss << typeKey(t.typeArgs[i]);
+                    }
+                    oss << ">";
+                }
+                return oss.str();
+            }
+            // Primitive / array labels
+            switch (t.kind) {
+                case Value::Type::Int:
+                    return "int";
+                case Value::Type::Float:
[…]
