+++ engineering/templates/adr-template.md
@@ -1,0 +1,15 @@
+# ADR-xxx: Title
+
+## Owner
+
+## Status
+
+## Delivered In
+
+## Context
+
+## Decision
+
+## Alternatives Considered
+
+## Consequences
