### 2026-01-29 #9625 @danielskatz
@bloch-akshay - it looks like you've merged the joss-paper-submission branch into the main branch. Is this correct? If so, I will change the branch that editorialbot uses to find the paper, and then continue the processing
