### 2026-01-29 #9625 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
