### 2026-01-29 #9625 @danielskatz
Congratulations to @bloch-akshay (Akshay Pal) on your JOSS publication!!

And thanks to @HectorMozo3110 and @joshiCoding for reviewing!
JOSS completely depends on volunteers - we couldn't do this without you
